% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cCOLOURS.R
\docType{data}
\name{cCOLORS}
\alias{cCOLORS}
\title{ANSI Colours for use in R (the "SuiteMFMR" selection)}
\format{
An object of class \code{environment} of length 32.
}
\usage{
cCOLORS
}
\description{
A collection of frequently used ANSI (American National Standards Institute)
Colours to support the "MFMR Suite of R Functions" (aka "SuiteMFMR").
}
\section{CONSTANTS}{

Functions for checking data integrity ...
}

\examples{
### Load the required R Library ...
library(MFMRutils)   # <- Loads library (if already installed locally) !!!

### ... the easily assign COLORS as follows ...
cCOLORS$BlackFORE     ### -> outputs the text in black font colour !!!
cCOLORS$CyanFORE      ### -> outputs the text in cyan font colour !!!
cCOLORS$MagentaBACK   ### -> outputs the text background in black font colour !!!

### Use with the direct-access R operator "::" from anywhere ...
MFMRutils::cCOLORS$CyanFORE   ### -> outputs the text in cyan font colour !!!

}
\keyword{datasets}
