% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cFORMATS.R
\docType{data}
\name{cFORMATS}
\alias{cFORMATS}
\title{ANSI Text Formats for use in R (the "SuiteMFMR" selection)}
\format{
An object of class \code{environment} of length 5.
}
\usage{
cFORMATS
}
\description{
A collection of frequently used ANSI (American National Standards Institute)
Text Font Formats to support the "MFMR Suite of R Functions" (aka "SuiteMFMR").
}
\examples{
### Load the required R Library ...
library(MFMRutils)   # <- Loads library (if already installed locally) !!!

### ... then easily apply TEXT FORMATS as follows ...
cFORMATS$BOLD      ### -> sets the text font format to BOLD !!!
cFORMATS$ITALICS   ### -> sets the text font format to ITALICS !!!
cFORMATS$RESET     ### -> removes actively applied ANSI Text Formatting !!!

### Use with the direct-access R operator "::" from anywhere ...
MFMRutils::cFORMATS$BOLD   ### -> sets the text font format to BOLD !!!

}
\keyword{datasets}
