% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cFSID.R
\docType{data}
\name{cFSID}
\alias{cFSID}
\title{The Func-Self-ID Args List (for use with "SuiteMFMR")}
\format{
An object of class \code{environment} of length 21.
}
\usage{
cFSID
}
\description{
A collection of frequently used ANSI (American National Standards Institute)
Text Font Formats to support the "MFMR Suite of R Functions" (aka "SuiteMFMR").
}
\examples{
### Load the required R Library ...
library(MFMRutils)   # <- Loads library (if already installed locally) !!!

### ... then easily apply TEXT FORMATS as follows ...
cFSID$PROJ_ID     ### -> sets the text font format to BOLD !!!
cFSID$FUNC_MODE   ### -> sets the text font format to ITALICS !!!
cFSID$TIME_STOP   ### -> removes actively applied ANSI Text Formatting !!!

### Use with the direct-access R operator "::" from anywhere ...
MFMRutils::cFSID$PROJ_ID   ### -> sets the text font format to BOLD !!!

}
\keyword{datasets}
