% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cICONS.R
\docType{data}
\name{cICONS}
\alias{cICONS}
\title{Icons for use in R (the "SuiteMFMR" selection)}
\format{
An object of class \code{environment} of length 33.
}
\usage{
cICONS
}
\description{
A collection of frequently used icons (in unicode form) \if{html}{\out{<originally>}} meant to
support the "MFMR Suite of R Functions" (aka "SuiteMFMR").
}
\examples{
### Load the required R Library ...
library(MFMRutils)   # <- Loads library (if already installed locally) !!!

### ... then easily assign ICONS as follows ...
cICONS$OoglyEyes      ### -> prints "👀" to the R console !!!
cICONS$CheckMark      ### -> prints "✔" to the R console !!!
cICONS$SmileyPonder   ### -> prints "🤔" to the R console !!!

### Use with the direct-access R operator "::" from anywhere ...
MFMRutils::cICONS$OoglyEyes   ### -> prints "👀" to the R console !!!

}
\keyword{datasets}
