% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cMFMR.R
\docType{data}
\name{cMFMR}
\alias{cMFMR}
\title{Full List of "SuiteMFMR" R Packages}
\format{
An object of class \code{environment} of length 7.
}
\usage{
cMFMR
}
\description{
The complete list of R Libraries that make up the "MFMR R Library Suite". This
Environment-Locked List contains the "standard" IDs \if{html}{\out{<tags>}} for the "MFMR Suite
of R Libraries" (aka "SuiteMFMR").
}
\examples{
### Load the required R Library ...
library(MFMRutils)   # <- Loads library (if already installed locally) !!!

### ... then easily extract the "SuiteMFMR" R Library IDs as follows ...
cMFMR$UTILS   # -> extracts the "Utils" MFMR R Library Identifier !!!
cMFMR$PLOTS   # -> extracts the "Plots" MFMR R Library Identifier !!!
cMFMR$STATS   # -> extracts the "Stats" MFMR R Library Identifier !!!

}
\keyword{datasets}
