% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CODE.Clean.File.Path.R
\name{code.clean.file.path}
\alias{code.clean.file.path}
\title{Sanitize String and Vector File Paths}
\usage{
code.clean.file.path(ssPathString = NULL, vsPathVector = NULL)
}
\arguments{
\item{ssPathString}{(\link{character}) A string value containing the File Path (i.e. local directory
text value) that needs to be sanitized (standardized) by this function.}

\item{vsPathVector}{(\link{vector} of \link{character}s) A character vector containing the File Path (i.e.
local directory text value) that needs to be sanitized (standardized) by
this function.}
}
\description{
Easily standardize R Project File Paths (directories) with the use of this "SuiteMFMR" function.
This function employs the base R file path function (i.e. \code{base::file.path()}) under the hood and
can handle R Project File Paths defined in the String and Vector formats/constructs.
}
\examples{
### Use the "Path-Cleaning" Function as follows: ...
library(MFMRutils)   # <- Loads the "MFMRutils" library (if already installed) ...



### OPTION 1 - Run a NULL-ARGs execution of the "Path-Cleaning" Function ...
# NB: If no external values are passed to the function a default file path will be returned !!!
ssPathArgsNULL <- code.clean.file.path()   # <- No external args <values> passed to function ... 
ssPathArgsNULL                             # -> returns a default R Project Path (".") !!!



### OPTION 2 - Use a STRING to define the File Path (R Project Directory, etc.) ...
# Prime the Function Inputs <arguments> (as needed) ...
ssCleanSTR_ <- "./rProjFiles/rData/TestDATA.csv"          # <- Clean PATH String !!!
ssMessySTR_ <- "./\\\\rProjFiles//rData/\\\\TestDATA.csv\\\\"   # <- Messy PATH String !!!

# Run the VECTOR-LOGIC execution of the "Path-Cleaning" Function ...
ssPathCleanSTRING <- MFMRutils::code.clean.file.path(   # <- Executes "STRING" Code Logic ...
  ssPathString = ssCleanSTR_                            # <- CLEAN "PATH String" supplied !!!
); ssPathCleanSTRING                                    # -> returns clean Path STRING result !!!

ssPathMessySTRING <- MFMRutils::code.clean.file.path(   # <- Executes the "STRING" Code Logic ...
  ssPathString = ssMessySTR_                            # <- MESSY "PATH String" supplied !!!
); ssPathMessySTRING                                    # -> returns clean Path STRING result !!!



### OPTION 3 - Use a VECTOR to define the File Path (R Project Directory, etc.) ...
# Prime the Function Inputs <arguments> (as needed) ...
vsCleanVEC_ <- c(".", "rProjFiles", "rData", "TestDATA.csv")             # <- Clean PATH Vector.
vsMessyVEC_ <- c("./", "\\\\rProjFiles/", "/rData/", "\\\\TestDATA.csv\\\\")   # <- Messy PATH Vector.

# Run the VECTOR-LOGIC execution of the "Path-Cleaning" Function ...
ssPathCleanVECTOR <- MFMRutils::code.clean.file.path(   # <- Executes the "VECTOR" Code Logic ...
  vsPathVector = vsCleanVEC_                            # <- CLEAN "PATH Vector" supplied !!!
); ssPathCleanVECTOR                                    # -> returns clean Path STRING result !!!

ssPathMessyVECTOR <- MFMRutils::code.clean.file.path(   # <- Executes the "VECTOR" Code Logic ...
  vsPathVector = vsMessyVEC_                            # <- MESSY "PATH Vector" supplied !!!
); ssPathMessyVECTOR                                    # -> returns clean Path STRING result !!!



### OPTION 4 - Provide BOTH the STRING and VECTOR Function Arguments ...
# NB: When both the String and Vector Function Arguments are supplied as function inputs, the
#     function will use the String argument instead of the Vector argument (i.e. defaults on the  
#     String function argument). 

# MESSY Example: run the "Path-Cleaning" Function with both the String & Vector MESSY values ...
ssPathBothArgsMESSY <- MFMRutils::code.clean.file.path(   # <- Executes "BOTH-ARGs" Code Logic...
  vsPathVector = vsMessyVEC_,                             # <- MESSY "PATH Vector" supplied !!!
  ssPathString = ssMessySTR_                              # <- MESSY "PATH String" supplied !!!
); ssPathBothArgsMESSY                                    # -> returns clean Path STRING result.

# CLEAN Example: run the "Path-Cleaning" Function with both the String & Vector CLEAN values ...
ssPathBothArgsCLEAN <- MFMRutils::code.clean.file.path(   # <- Executes "BOTH-ARGs" Code Logic...
  vsPathVector = vsCleanVEC_,                             # <- CLEAN "PATH Vector" supplied !!!
  ssPathString = ssCleanSTR_                              # <- CLEAN "PATH String" supplied !!!
); ssPathBothArgsCLEAN                                    # -> returns clean Path STRING result.



}
