% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CODE.Tiny.Helpers.EXPORTED.R
\name{code.coalesce.dots.args}
\alias{code.coalesce.dots.args}
\title{Dots-Args Coalescing Function}
\usage{
code.coalesce.dots.args(coArgsList01 = NULL, coArgsList02 = NULL)
}
\arguments{
\item{coArgsList01}{(\link{list}) an R List (complex) object that captures the first group of Function
Dot Arguments (i.e. "Fall-Through" Function Arguments) to be coalesced.}

\item{coArgsList02}{(\link{list}) an R List (complex) object that captures the second group of
Function Dot Arguments (i.e. "Fall-Through" Function Arguments) to be
coalesced.}
}
\description{
A Function Dots Arguments Coalescing Function. This function greatly aides the unification (i.e.
joining, collapsing or coalescing) of two sets of "DotsArgs" groups into a single R List Object.
This is a very important function within the "MFMR Suite of R Functions" (i.e. \code{SuiteMFMR})
considering almost all \code{SuiteMFMR} functions implement a dual-approach (i.e. more user-friendly
approach) towards passing around "Fall-Through" (i.e. "DotsArgs") function arguments.
}
\examples{
### Use the DotsArgs-Coalescing Function as follows: ...
library(MFMRutils)   # <- Ensures the "MFMRutils" library is installed locally (loads library)...

### Then apply the NCO accordingly ...


}
\seealso{
Other CODE Functions (SuiteMFMR): 
\code{\link{\%??\%}()},
\code{\link{code.poll.r.runtime.mode}()},
\code{\link{code.return.renv.list}()}
}
\concept{CODE Functions (SuiteMFMR)}
