% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CODE.Find.Text.Instances.R
\name{code.find.text.instances}
\alias{code.find.text.instances}
\title{Locate Text or Regex Strings in R Functions}
\usage{
code.find.text.instances(
  csFindText = "unc_this",
  ssSearchFunc = NULL,
  coRENV = .GlobalEnv,
  sbRetFullMatch = TRUE
)
}
\arguments{
\item{csFindText}{(\link{character}) A string value containing the text or regular expression (regex)
value or pattern to match against (search for) in the R Function code base.}

\item{ssSearchFunc}{(\link{character}) A string value containing the function name identifier of the
R Function where the search should be conducted or targeted at. This is an
optional argument - which means the function will search all functions within
the active Global R Environment (\link{.GlobalEnv}) if this function argument is
set to a value of \link{NULL}.}

\item{coRENV}{(\link{environment}) An R Object argument that defines the scope (i.e. R Environment)
where the text/regex search should be conducted (default: Global R Environment).}

\item{sbRetFullMatch}{(\link{logical}) A boolean function that defines whether the full matching line
(i.e. text/string or code snippet) should be returned as a function output.
If set to \link{FALSE} then only the text/regex provided to the function via
the \code{csFindText} function argument will be returned. If set to \link{TRUE} then
all \if{html}{\out{<contiguous>}} text on the matching \if{html}{\out{<code>}} line will be returned.}
}
\value{
This function returns a Data Frame that contains the following variables \if{html}{\out{<columns>}}:
\itemize{
\item FUNC_NAME   -> a \link{character} identifier of the R Function where the matching regex was found.
\item LINE_NUMBER -> an \link{integer} variable denoting the R Function code line number where the
matching text or regex value was found.
\item CODE_SNIP   -> a \link{character} variable that captures a code snippet from the matching line where
the text or regex value was found.
\item FILE_NAME   -> a \link{character} variable that captures the \if{html}{\out{<parent>}} file identifier (i.e. name) of
the R Function where the matching text or regex value was found.
}
}
\description{
Easily search and find text or regular expression (regex) snippets in any R Function.
}
\examples{
### Use the "Path-Cleaning" Function as follows: ...
library(MFMRutils)   # <- Loads the "MFMRutils" library (if already installed) ...



### OPTION 1 - Run a NULL-ARGs execution of the "Path-Cleaning" Function ...
# NB: If no external values are passed to the function a default file path will be returned !!!
ssPathArgsNULL <- code.clean.file.path()   # <- No external args <values> passed to function ... 
ssPathArgsNULL                             # -> returns a default R Project Path (".") !!!



### OPTION 4 - Provide BOTH the STRING and VECTOR Function Arguments ...
# NB: When both the String and Vector Function Arguments are supplied as function inputs, the
#     function will use the String argument instead of the Vector argument (i.e. defaults on the  
#     String function argument). 

# MESSY Example: run the "Path-Cleaning" Function with both the String & Vector MESSY values ...
ssPathBothArgsMESSY <- MFMRutils::code.clean.file.path(   # <- Executes "BOTH-ARGs" Code Logic...
  vsPathVector = vsMessyVEC_,                             # <- MESSY "PATH Vector" supplied !!!
  ssPathString = ssMessySTR_                              # <- MESSY "PATH String" supplied !!!
); ssPathBothArgsMESSY                                    # -> returns clean Path STRING result.

# CLEAN Example: run the "Path-Cleaning" Function with both the String & Vector CLEAN values ...
ssPathBothArgsCLEAN <- MFMRutils::code.clean.file.path(   # <- Executes "BOTH-ARGs" Code Logic...
  vsPathVector = vsCleanVEC_,                             # <- CLEAN "PATH Vector" supplied !!!
  ssPathString = ssCleanSTR_                              # <- CLEAN "PATH String" supplied !!!
); ssPathBothArgsCLEAN                                    # -> returns clean Path STRING result.



}
