% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CODE.Get.CELN.R
\name{code.get.celn}
\alias{code.get.celn}
\title{Extract the Code Editor Gutter Line Number (CELN)}
\usage{
code.get.celn(
  ssFuncName = "code.get.celn\\\\(",
  sbRunByForce = FALSE,
  siCallIndex = 1L,
  sbUseAlias = FALSE,
  ssAliasValue = "rasMfmrGetCELN\\\\("
)
}
\arguments{
\item{ssFuncName}{(\link{character}) a String or Text (i.e. \link{character} \link{vector}) value that specifies
the Function Identifier (i.e. name) where the R Code Search or Function ID
look-up should be conducted (i.e. directed at).}

\item{sbRunByForce}{(\link{logical}) a Boolean value that specifies whether the function must execute
its standard code logic even if the R Project Development DEBUG tracker (
\code{RCT_IS_DEBUG_RT_MODE_}) was activated (i.e. found in the call stack & set to
a value of \code{TRUE}) or not (default: \code{FALSE}).}

\item{siCallIndex}{(\link{integer}) an Integer value that defines the "Call Index" or "Call Sequence"
in the localized (function \if{html}{\out{<internal>}}) call stack in reference to when (i.e.
in which order) the \code{code.get.celn()} function was called (i.e. from inside a
function).}

\item{sbUseAlias}{(\link{logical}) a Boolean value that specifies whether the function should use the
the standard \code{code.get.celn()} function identifier or not (default: \code{FALSE}).
If set to \code{TRUE} the function will NOT use standard function ID (i.e.
"code\\.get\\.celn"), but rather the value supplied to the function via the
\code{ssAliasValue} function argument.}

\item{ssAliasValue}{(\link{character}) a String or Text (i.e. \link{character} \link{vector}) value that
specifies the alias (i.e. re-assigned function identifier) of the default or
standard \code{code.get.celn()} function identifier to search for in the R
function \if{html}{\out{<internal>}} code.}
}
\description{
A SPECIAL HELPER Function that extracts the Code Editor Gutter Line Number (CEGLN a.k.a. CELN) at
the code editor line location from which the \code{code.get.celn()} function was called. This function
was created for the purpose of enhancing the debugging, R Package development and real-time (i.e.
interactive) code documentation processes from within the MFMR Suite of R Functions. However,
this function may also be used as a stand-alone function by (or in) other, 3rd Party, libraries
and custom R functions - as long as the \code{MFMRutils} library is linked (i.e. previously installed
and/or specified as a library dependency).
}
\examples{
### Easily debug custom R function code with this <cool> helper function ...
library(MFMRutils)   # <- Ensures the `MFMRutils` library is <already> installed locally.



### Use with any custom R function as follows ...
"my.custom.r.func" <- function(x=7, y=3, z=28) {

  RCT_TAG_FUNC_ID <- "my.custom.r.func";  # <- ALWAYS TAG (ID) Custom R Functions accordingly !!!

  valSUM <- sum(x, y, z);
  cat(
    paste0(
      " \u279C ", RCT_TAG_FUNC_ID, " " , 
      MFMRutils::code.get.celn(
        RCT_TAG_FUNC_ID,   # <- Provide the Function ID (tag) for the Custom R Function ...
        TRUE,              # <- <force> Run function even if DEBUG tracking is not activated !!!
        1L                 # <- Define the call order <call sequence> for the `CODE.Get.CELN()` 
      ),                   #    function calls inside the custom function code. This is the 1st 
                           #    time the `CODE.Get.CELN()` function is called inside this custom 
                           #    function so the `siCallIndex` value here should == `1L`.
      " | Summed all 3 input values < result: ", valSUM," > !!! \n"
    )
  );

  valMEAN <- sum(x, y, z) / 3;
  cat(
    paste0(
      " \u279C ", RCT_TAG_FUNC_ID, " " ,
      MFMRutils::code.get.celn(
        RCT_TAG_FUNC_ID,   # <- Provide the Function ID (tag) for the Custom R Function ...
        TRUE,              # <- <force> Run function even if DEBUG tracking is not activated !!!
        2L                 # <- Define the call order <call sequence> for the `CODE.Get.CELN()` 
      ),                   #    function calls inside the custom function code. This is the 2nd 
                           #    time the `CODE.Get.CELN()` function is called inside this custom 
                           #    function so the `siCallIndex` value here should == `2L`.
      " | Averaged the 3 input values < result: ", round(valMEAN, 1)," > !!! \n"
    )
  );
  
  # Notify function's SUCCESSFUL COMPLETION ( <- is only needed for code illustration) ...
  cat(
    paste0(
      " \u279C ", RCT_TAG_FUNC_ID, " " ,
      MFMRutils::code.get.celn(
        RCT_TAG_FUNC_ID,   # <- Provide the Function ID (tag) for the Custom R Function ...
        TRUE,              # <- <force> Run function even if DEBUG tracking is not activated !!!
        3L                 # <- Define the call order <call sequence> for the `CODE.Get.CELN()` 
      ),                   #    function calls inside the custom function code. This is the 3rd 
                           #    time the `CODE.Get.CELN()` function is called inside this custom 
                           #    function so the `siCallIndex` value here should == `3L`.
      " | CUSTOM R-FUNC `", tolower(RCT_TAG_FUNC_ID), "()` SUCESSFULLY EXECUTED !!! \n\n"
    )
  );

  return(
    list("SUM" = valSUM, "MEAN" = valMEAN)
  );
}

## Execute the custom R Function ...
my.custom.r.func()

## Outputs from "my.custom.r.func()" should be as follows ...
# ➜ My.Cust.R.FUNC 9 | Summed all 3 input values < result: 38 > !!!
# ➜ My.Cust.R.FUNC 24 | Averaged the 3 input values < result: 12.7 > !!!
# ➜ My.Cust.R.FUNC 39 | CUSTOM R-FUNC `my.custom.r.func()` SUCESSFULLY EXECUTED !!!

# $SUM
# [1] 38

# $MEAN
# [1] 12.66667


}
\concept{CODE Functions (\code{SuiteMFMR})}
