% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.get.egln.R
\name{code.get.egln}
\alias{code.get.egln}
\title{Extract the RStudio Editor Gutter Line Number ...}
\usage{
code.get.egln()   ### <- if "MFMRutils" library is already installed & loaded !!!
}
\description{
A SPECIAL HELPER Function that extracts the Editor Gutter Line Number - EGLN (at the line location from which this function was called). This function only works in the RStudio IDE since it makes direct use of the "rstudioapi" library. This function was created with the main purpose of enhancing the debugging and R Package development process from within the MFMR Suite of R Functions.
}
\examples{
### Easily debug your R code with the help of this <special> function ...
code.get.egln()   ### <- prints the active script editor line number to the R console !!!

### Use with the direct-access R operator "::" from anywhere ...
MFMRutils::code.get.egln()   ### <- prints the active script editor line number to the R console !!!

}
