% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CODE.Return.Env.Locked.List.R
\name{code.return.env.locked.list}
\alias{code.return.env.locked.list}
\title{Compile Environment Locked R Lists}
\usage{
code.return.env.locked.list(vsListNames, lsListVals, sbLockList = FALSE)
}
\arguments{
\item{vsListNames}{a vector of characters (strings) that captures the "names" of the environment
locked list. The "vsListNames" function argument must match the "lsListVals"
function argument in terms of length (i.e. number of vector elements), as well
as the required (correct) NAME-VALUE pairing (i.e. NAME-VALUE pairs).}

\item{lsListVals}{a list of R Objects that contains the "values" of the environment locked list.
The "lsListVals" function argument must match the "vsListNames" function
argument in terms of length (i.e. number of vector elements), as well as the
required (correct) NAME-VALUE pairing (i.e. NAME-VALUE pairs).}

\item{sbLockList}{a logical (boolean) value that denotes whether the returned \if{html}{\out{<output>}} list
should be R environment locked <Default == FALSE>. If TRUE, individual list
bindings (values) are locked, preventing modification of existing list elements.}
}
\value{
\itemize{
\item This function returns a locked environment R Object.
}
}
\description{
A relatively small R function that compiles and returns R Environment Locked lists.
}
\examples{
### Activate the "MFMRutils" R Library (if previously installed) ...
library(MFMRutils)   ### -> Loads the "MFMRutils" library !!!

### Compile the R Locked List Information (for active R Library <project>) ...
vsListNames_ <- c("VAR_A", "VAR_B", "VAR_C", "VAR_X", "VAR_Y", "VAR_Z", "VAR_G")
lsListVals_  <- list(
  1982, "Value for VAR_B", "R-Object for VAR_C", FALSE, 
  "Value for VAR_Y", TRUE, "R-List for VAR_G"
)



### Function-use OPTION 1 (main purpose) -> Create Immutable R List Objects ... 
rlsEnvLockdLIST <- code.return.env.locked.list(
  vsListNames = vsListNames_, lsListVals = lsListVals_, 
  sbLockList = TRUE   # <- Set to 'TRUE' to create an immutable (environment locked) R List !!!
)

rlsEnvLockdLIST$VAR_X        # -> Returns the value FALSE !!!
rlsEnvLockdLIST[["VAR_C"]]   # -> Returns the value 'R-Object for VAR_C' !!! 
rlsEnvLockdLIST$VAR_G        # -> Returns the value 'R-List for VAR_G' !!! 

\dontrun{   ### <- Code below does an "immutability" check (to ensure the custom function works
            ###    as intended) -> but this code creates a variable write error under the normal
            ###    "R_CMD_CHECK" scenario and as such should not be run during normal code check.
## Immutability test (OPTION 1 test) ...
rlsEnvLockdLIST$VAR_G <- "A NEW value for 'VAR_G' !!!"   # -> Will trigger an error !!!

}


 
### Function-use OPTION 2 (secondary purpose) -> Create Mutable R List Objects ... 
rlsEnvLockdLIST <- code.return.env.locked.list(
  vsListNames = vsListNames_, lsListVals = lsListVals_, 
  sbLockList = FALSE   # <- Set to 'FALSE' to create a mutable R List !!!
)

## Mutability test (OPTION 2 test) ...
rlsEnvLockdLIST$VAR_G <- "A NEW value for 'VAR_G' !!!"   # -> Assigns a new value to the "VAR_G"
                                                         #    element (name) of the R list.

}
