% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CODE.Return.RENV.List.R
\name{code.return.renv.list}
\alias{code.return.renv.list}
\title{Compile R-Environment Locked Lists}
\usage{
code.return.renv.list(
  vsListNames = NULL,
  lsListVals = NULL,
  sbLockList = FALSE,
  sbSetClass = FALSE,
  ssClassValue = "RENV-LOCKD-LIST",
  ...
)
}
\arguments{
\item{vsListNames}{(\link{vector} of \link{character}s) a Vector of String or Text (\link{character}) values
that contain the "names" or "keys" of the R Environment Locked List to be
compiled & returned. VERY NB: The \code{vsListNames} function argument must match
the \code{lsListVals} function argument in terms of length (i.e. number of vector
elements) and order (i.e. sequence in accordance to the required NAME-VALUE
or Key-Value pairing).}

\item{lsListVals}{(\link{list}) a List of R Objects that captures the "values" of the R Environment
Locked List to be compiled & returned. VERY NB: The \code{lsListVals} function
argument must match the \code{vsListNames} function argument in terms of length (
i.e. number of vector elements) and order (i.e. sequence in accordance to the
required NAME-VALUE or Key-Value pairing).}

\item{sbLockList}{(\link{logical}) a Boolean value that defines whether the returned \if{html}{\out{<output>}} list
should be R Environment Locked or not (default: \code{FALSE}). If TRUE, individual
list bindings (i.e. Key-Value Pairs) are locked, preventing modification of
existing list elements -> making the returned \if{html}{\out{<output>}} list immutable.}

\item{sbSetClass}{(\link{logical}) a Boolean value that denotes whether the class of the returned
\if{html}{\out{<output>}} list should be left as the default R type (i.e. \link{environment}) or
not (default: \code{FALSE}). If set to \code{TRUE} the class of the output list will be
set to the value of the \code{ssClassValue} function argument.}

\item{ssClassValue}{(\link{character}) a String or Text (i.e. \link{character} \link{vector}) value that
captures the \if{html}{\out{<custom>}} class value of the returned \if{html}{\out{<output>}} R Environment
Locked List.}

\item{...}{(\link{list}) a List of R Objects function argument to catch all "Fall-Through" or
"DotsArgs" values for functions that are nested within the \code{\link[=code.return.renv.list]{code.return.renv.list()}}
function. The \link{...} argument here has specific relevance (reference) to the following
nested functions: \if{html}{\out{<br>}}
1. \code{\link[=info.post.func.self.id]{info.post.func.self.id()}} , \if{html}{\out{<br>}}
2. \code{\link[=info.post.note]{info.post.note()}} , and \if{html}{\out{<br>}}
3. \code{\link[=code.get.celn]{code.get.celn()}} .}
}
\value{
\itemize{
\item This function returns a locked environment R Object.
}
}
\description{
A relatively small R function that compiles and returns R Environment Locked lists.
}
\examples{
### Activate the "MFMRutils" R Library (if previously installed) ...
library(MFMRutils)   ### -> Loads the "MFMRutils" library !!!

### Compile the R Locked List Information (for active R Library <project>) ...
vsListNames_ <- c("VAR_A", "VAR_B", "VAR_C", "VAR_X", "VAR_Y", "VAR_Z", "VAR_G")
lsListVals_  <- list(
  1982, "Value for VAR_B", "R-Object for VAR_C", FALSE, 
  "Value for VAR_Y", TRUE, "R-List for VAR_G"
)



### Function-use OPTION 1 (main purpose) -> Create Immutable R List Objects ... 
rlsEnvLockdLIST <- code.return.renv.list(
  vsListNames = vsListNames_, lsListVals = lsListVals_, 
  sbLockList = TRUE   # <- Set to 'TRUE' to create an immutable (environment locked) R List !!!
)

rlsEnvLockdLIST$VAR_X        # -> Returns the value FALSE !!!
rlsEnvLockdLIST[["VAR_C"]]   # -> Returns the value 'R-Object for VAR_C' !!! 
rlsEnvLockdLIST$VAR_G        # -> Returns the value 'R-List for VAR_G' !!! 

\dontrun{   ### <- Code below does an "immutability" check (to ensure the custom function works
            ###    as intended) -> but this code creates a variable write error under the normal
            ###    "R_CMD_CHECK" scenario and as such should not be run during normal R library
            ###    development code check procedures.
## Immutability test (OPTION 1 test) ...
rlsEnvLockdLIST$VAR_G <- "A NEW value for 'VAR_G' !!!"   # -> Will trigger an error !!!

}


 
### Function-use OPTION 2 (secondary purpose) -> Create Mutable R List Objects ... 
rlsEnvLockdLIST <- code.return.renv.list(
  vsListNames = vsListNames_, lsListVals = lsListVals_, 
  sbLockList = FALSE   # <- Set to 'FALSE' to create a mutable R List !!!
)

## Mutability test (OPTION 2 test) ...
rlsEnvLockdLIST$VAR_G <- "A NEW value for 'VAR_G' !!!"   # -> Assigns a new value to the "VAR_G"
                                                         #    element (name) of the R list.

rlsEnvLockdLIST$VAR_G        # -> Returns the value "A NEW value for 'VAR_G' !!!" !!!


}
\seealso{
Other CODE Functions (SuiteMFMR): 
\code{\link{\%??\%}()},
\code{\link{code.coalesce.dots.args}()},
\code{\link{code.poll.r.runtime.mode}()}
}
\concept{CODE Functions (SuiteMFMR)}
