% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CODE.Split.String.On.Vector.R
\name{code.split.string.on.vector}
\alias{code.split.string.on.vector}
\title{Split String on Basis of Vector Elements}
\usage{
code.split.string.on.vector(
  ssSplitString = NULL,
  vsSplitVector = NULL,
  sbExactSplits = FALSE
)
}
\arguments{
\item{ssSplitString}{(\link{character}) A function argument that supplies the string or text (to be
cut up into its constituent parts) to the internal code of this function.}

\item{vsSplitVector}{(\link{vector} of \link{character}s) A vector of characters (strings) that defines the
characters (symbols) at which the \code{ssSplitString} value should be cut up at.}

\item{sbExactSplits}{(\link{logical}) A boolean function argument that defines the type of split to
apply (default: \link{FALSE}). If \link{TRUE} it interprets the \code{vsSplitVector} values
exactly (i.e. as is -> on their base or literal character symbolization), or
else the function applies the \code{vsSplitVector} values as regular expressions.}
}
\description{
Effortlessly cut up (split) an R String (i.e. an R character vector) into the required pieces
(snippets) with the help of a second character vector that specifies the locations (i.e. symbols
or character values) where the main character vector (String) should be cut up at.
}
\examples{
### Use the "Vector-Splitting" Function as follows: ...
library(MFMRutils)   # <- Loads the "MFMRutils" library (if already installed) ...


### Prime the Function Inputs (accordingly) ...
ssSplitString_ <- base::paste0(   # <- Compile a SUPER MESSED-UP <symbol-riddled> String.
"A+ ", " -", "CRAZY&", " // ", "+ String", " ?\%? ", "to", "_ ", "& test", " \\\\ ", "} the", 
"/ ", "& SUPER{-\%-}COOL#", "?!", "FUNCTIONALITY", " @ ", "of", " # ", "the {", "? ???&???", 
"+ 'MFMRutils::code.split.string.on.vector()'", " $ ", "^ + function", "!! !"
); ssSplitString_   # <- Placing the variable here (after ";") merely prints it to the R Console.

vsSplitVector_ = c(   # <- Specify on which characters the MESSED-UP String should be cut up at.
  "/", "_", "-", "?", " ", "!", "'", "\\\\", "@", "#", "^", "$", "+", "\%", "}", "&", "{"
); vsSplitVector_   # <- Placing the variable here (after ";") merely prints it to the R Console.


### Then execute the "Vector-Splitting" Function ...
rlsResSplit <- code.split.string.on.vector(   # <- Runs the actual String Split process ...
  ssSplitString = ssSplitString_,
  vsSplitVector = vsSplitVector_, 
)


### Finally -> view the results of the "Vector-Splitting" Function ...
rlsResSplit$SPLITS   # -> Returns the vector of String Splits (i.e. all parts of split string). 
rlsResSplit$STRING   # -> Returns a clean string version of the String Splits (splits combined).


}
