% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CODE.Split.String.Via.Vector.R
\name{code.split.string.via.vector}
\alias{code.split.string.via.vector}
\title{Sanitize String and Vector File Paths}
\usage{
code.split.string.via.vector(
  ssStrToSplit =
    "A-CRAZY//TEST\\\\String?for the?!'MFMRutils::split.string.via.vector()'_function!",
  vsSplitVector = c("/", "_", "-", "?", "\\\\", " ", "'", "!"),
  sbExactSplits = FALSE
)
}
\arguments{
\item{ssStrToSplit}{a dynamic (complex) object that captures the "Left-Hand" R Object
to be evaluated against the "NULL Criteria" (i.e. by way of value,
length or class \if{html}{\out{<data type>}} NULL Checks).}

\item{vsSplitVector}{a dynamic (complex) object that captures the "Right-Hand" R Object
(to be returned if the "Left-Hand" R Object is NULL).}

\item{sbExactSplits}{\link{logical} If TRUE match split exactly, otherwise use regular expressions.}
}
\description{
The "SuiteMFMR" Null-Coalescing Operator (NCO) is similar to the "??" NCO of
the DART Programming Language. The operator evaluates whether the "Left-Hand" R
Object is NULL (in terms of its value or length or class \if{html}{\out{<data type>}}) and returns
the "Right-Hand" R Object if the "Left-Hand" Object is \if{html}{\out{<indeed>}} NULL.
}
\examples{
### Use the Null-Coalescing Operator (NCO) as follows: ...
library(MFMRutils)   # <- Loads the "MFMRutils" library (if already installed) ...

### Then apply the NCO accordingly ...
# Use Strings as File Paths (i.e. as R Project Directory Inputs) ...


# Use Vectors as File Paths (i.e. as R Project Directory Inputs) ...
vsPathCLEAN_ <- c(".", "rProjFiles", "rData", "InputDATA.txt");            # <- CLEAN Vector
vsPathMESSY_ <- c("./", "\\\\rProjFiles/", "/rData/", "\\\\TestDATA.txt\\\\");   # <- MESSY Vector

}
