% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEVS.Create.R.Function.R
\name{devs.create.r.function}
\alias{devs.create.r.function}
\title{CRAN Code Validation with Code Version Tracking ("SuiteMFMR" DevTools)}
\usage{
devs.create.r.function(
  ssFuncID = "CODE.New.R.FUNC",
  ssFuncDir = "./R",
  sbIsLibrFunc = TRUE,
  sbForceFuncPath = FALSE,
  sbForceFuncCreate = FALSE,
  ...
)
}
\arguments{
\item{...}{the fall-through function arguments (i.e. DotsArgs) used for nested functions within
this main function. The "DotsArgs" implementation here pertain to the following nested
R functions (i.e. R functions used within this main R function): \if{html}{\out{<br>}}
1. \code{\link[=info.post.func.self.id]{info.post.func.self.id()}}, and \if{html}{\out{<br>}}
2. \code{\link[=devs.patch.code.dev.trckr.file]{devs.patch.code.dev.trckr.file()}}.}

\item{sbCheckDocs}{a logical (boolean) value that specifies whether to run the standard package
documentation process only (not the FULL CRAN Code Validation Process/Check).}

\item{sbCheckCRAN}{a logical (boolean) value that specifies whether to run the FULL CRAN Code
Validation Process/Check (inclusive of the pre-check documentation steps).}

\item{ssTimeZone}{a simple character vector (string) that defines the Time Zone to be used for
the package documentation.}

\item{sbIsProdRel}{a logical (boolean) value that captures if the code-check and code commit
process (action) is a "Production Release" code development action or not.}
}
\value{
\itemize{
\item This function returns the programmatically amended or updated (real-time or active) version
number for the active R Library Project as a list of character objects.
\item This function creates a Work-In-Progress (WIP) directory at the root of the active R-Libs
Project (i.e. "./WIP" <- if not already exists). The "./WIP" folder and all of its contents
are included in GIT push-&-pull processes. You must MANUALLY ADD the following code stub
(^WIP$) \if{html}{\out{<sans parentheses>}} to the ".Rbuildignore" R Project file to ensure the "./WIP"
folder + contents are excluded from the R Project Build Process. (<- this is IMPORTANT !!!)
\item This function also creates a "DevsVersTimeStamp.txt" file in the "./WIP" folder for secondary
development version tracking purposes (as needed).
}
}
\description{
A Helper Function that executes the CRAN pre-requisite Code Checking Procedure during active R
Package Development. This function programmatically updates the package version number in the R
Project DESCRIPTION file before running the required documentation and CRAN Package
Pre-Submission Code Requirement Checks (i.e. CRAN Code Validation) during iterative development
cycles. This Helper Function was originally crafted as technical support tool for the "MFMR
Suite of R Libraries" (aka "SuiteMFMR"), but may also be helpful as a support tool during the R
Package Development Cycle of other (3rd Party) R Projects.
}
\examples{
\dontrun{   ### <- This function constitutes a development utility !!! This function creates (and
            ###    requires) a special development directory ("./WIP") that is created during the
            ###    init-run (initial R project setup) phase and is intended to facilitate a user-
            ###    friendly R Library development process. For these reasons the code examples 
            ###    below should not be executed during "R_CMD_CHECK" code check procedures.
            
### Run R Package DevCode easily as follows ...
library(MFMRutils)   # <- Loads "MFMRutils" library (if already installed) !!!

### Run 2 different types of code check/validation processes ...
devs.check.code.specs(sbCheckDocs = TRUE)   # -> Executes only the DevTools Documentation 
                                            #    Process.
devs.check.code.specs(sbCheckCRAN = TRUE)   # -> Executes the more complete (stringent) CRAN Code 
                                            #    Validation Process.

### Check (i.e. "rasDevToolsCHECK()") overrides the documentation process ...
# The complete (CRAN) Documentation Process will only be executed once if both are TRUE !!!
devs.check.code.specs(sbCheckDocs = TRUE, sbCheckCRAN = TRUE)

}

}
\seealso{
Other SuiteMFMR DevTools: 
\code{\link{devs.check.code.specs}()},
\code{\link{devs.find.code.instances}()},
\code{\link{devs.patch.code.dev.trckr.file}()},
\code{\link{devs.pull.libr.info}()}
}
\concept{SuiteMFMR DevTools}
