% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENVColors.R
\docType{data}
\name{ENVColors}
\alias{ENVColors}
\title{ANSI Colours for use in R (the \code{MFMRutils} selection)}
\format{
An object of class \code{environment} of length 32.
}
\usage{
ENVColors   ### -> if [MFMRutils] is already installed & loaded !!!
}
\description{
A collection of frequently used ANSI (American National Standards Institute)
Colours to support the MFMR Suite of R Functions.
}
\examples{
### Easily print & assign icons as follows ...
ENVColors$BlackFORE     ### -> outputs the text in black font colour !!!
ENVColors$CyanFORE      ### -> outputs the text in cyan font colour !!!
ENVColors$MagentaBACK   ### -> outputs the text background in black font colour !!!

### Use with the direct-access R operator "::" from anywhere ...
MFMRutils::ENVColors$CyanFORE   ### -> outputs the text in cyan font colour !!!

}
\keyword{datasets}
