% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvDATEs.R
\docType{data}
\name{EnvDATES}
\alias{EnvDATES}
\title{Date-Time Formats for use in R (the "SuiteMFMR" selection)}
\format{
An object of class \code{environment} of length 6.
}
\usage{
EnvDATES
}
\description{
A collection of frequently used Date and Date-Time (text) Formats to support
the "MFMR Suite of R Functions" (aka "SuiteMFMR").
}
\examples{
### Load the required R Library ...
library(MFMRutils)   # <- Loads library (if already installed locally) !!!

### ... then easily assign DATE FORMATS as follows ...
EnvDATES$SHORTv01   ### -> formats Date to "Saturday, January 16 1982" !!!
EnvDATES$LONGv02    ### -> formats Date to "Sat, Jan 16 1982 @ 21:57:38" !!!
EnvDATES$LONGv03    ### -> formats Date to "Sat, 16 Jan 1982 @ 21:57:38" !!!
EnvDATES$SHORTv03   ### -> formats Date to "Sat, 16 Jan 1982" !!!

### Use with the direct-access R operator "::" from anywhere ...
MFMRutils::EnvDATES$SHORTv02   ### -> formats Date to "Sat, Jan 16 1982" !!!

}
\keyword{datasets}
