% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvArgsFSID.R
\docType{data}
\name{EnvFSID}
\alias{EnvFSID}
\title{The Func-Self-ID Args List (for use with "SuiteMFMR")}
\format{
An object of class \code{environment} of length 19.
}
\usage{
EnvFSID
}
\description{
A collection of frequently used ANSI (American National Standards Institute)
Text Font Formats to support the "MFMR Suite of R Functions" (aka "SuiteMFMR").
}
\examples{
### Load the required R Library ...
library(MFMRutils)   # <- Loads library (if already installed locally) !!!

### ... then easily apply TEXT FORMATS as follows ...
EnvFORMATS$BOLD      ### -> sets the text font format to BOLD !!!
EnvFORMATS$ITALICS   ### -> sets the text font format to ITALICS !!!
EnvFORMATS$RESET     ### -> removes actively applied ANSI Text Formatting !!!

### Use with the direct-access R operator "::" from anywhere ...
MFMRutils::EnvFORMATS$BOLD   ### -> sets the text font format to BOLD !!!

}
\keyword{datasets}
