% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENVIcons.R
\docType{data}
\name{ENVIcons}
\alias{ENVIcons}
\title{Unicode Icons for use in R (the \code{MFMRutils} selection)}
\format{
An object of class \code{environment} of length 9.
}
\usage{
ENVIcons   ### -> if "MFMRutils" library is already installed & loaded !!!
}
\description{
A collection of frequently used icons (Unicode characters) meant to support
the MFMR Suite of R Functions.
}
\examples{
### Easily print & assign icons as follows ...
ENVIcons$OoglyEyes      ### -> prints "ğŸ‘€" to the R console !!!
ENVIcons$CheckMark      ### -> prints "âœ”" to the R console !!!
ENVIcons$SmileyPonder   ### -> prints "ğŸ¤”" to the R console !!!

### Use with the direct-access R operator "::" from anywhere ...
MFMRutils::ENVIcons$OoglyEyes   ### -> prints "ğŸ‘€" to the R console !!!

}
\keyword{datasets}
