% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvLibsMFMR.R
\docType{data}
\name{EnvLibsMFMR}
\alias{EnvLibsMFMR}
\title{The MFMR Suite of R Libraries ...}
\format{
An object of class \code{environment} of length 6.
}
\usage{
EnvLibsMFMR
}
\description{
The complete list of R Libraries that make up the "MFMR R Library Suite". This
Environment-Locked List contains the "standard" IDs for the MFMR R Libraries.
}
\examples{
### Easily print & assign Date Formats as follows ...
ENVLibsMFMR$UTILS   ### -> formats Date to "Saturday, January 16 1982" !!!
ENVLibsMFMR$UTILS   ### -> formats Date to "Sat, Jan 16 1982" !!!
ENVLibsMFMR$UTILS   ### -> formats Date to "Sat, 16 Jan 1982" !!!

}
\keyword{datasets}
