% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvMiscCONSTs.R
\docType{data}
\name{EnvMiscCONSTs}
\alias{EnvMiscCONSTs}
\title{The "SuiteMFMR" List of Useful Constants}
\format{
An object of class \code{environment} of length 6.
}
\usage{
EnvMiscCONSTs
}
\description{
A collection of surprisingly useful CONSTANTS that are required throughout the
"SuiteMFMR" R Project. This Environment-LOCKED R List standardizes & improves
the extraction & use of key values across multiple functions of the "MFMR Suite
of R Libraries (aka "SuiteMFMR").
}
\examples{
### Load the required R Library ...
library(MFMRutils)   # <- Loads library (if already installed locally) !!!

### ... then easily apply TEXT FORMATS as follows ...
EnvMiscCONSTs$PATH_FOLDER_WIP_PROD   # -> Returns the Work-In-Progress (WIP) 
                                     #    path for the R Library Development
                                     #    Staging Folder ("01_NextForPROD") !!!

### Use with the direct-access R operator "::" from anywhere ...
MFMRutils::EnvMiscCONSTs$PATH_FOLDER_WIP_PROD

}
\keyword{datasets}
