% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiny.lists.and.enums.R
\docType{data}
\name{IconsLUCCs}
\alias{IconsLUCCs}
\title{Unicode Icons for use in R (the \code{MFMRutils} selection)}
\format{
An object of class \code{environment} of length 8.
}
\usage{
IconsLUCCs   ### <- if "MFMRutils" library is already installed & loaded !!!
}
\description{
A collection of frequently used icons (Unicode characters) meant to support
the MFMR Suite of R Functions.
}
\examples{
### Easily print & assign icons as follows ...
IconsLUCCs$OoglyEyes      ### <- prints "👀" to the R console !!!
IconsLUCCs$CheckMark      ### <- prints "✔" to the R console !!!
IconsLUCCs$SmileyPonder   ### <- prints "🤔" to the R console !!!

### Use with the direct-access R operator "::" from anywhere ...
MFMRutils::IconsLUCCs$OoglyEyes   ### <- prints "👀" to the R console !!!

}
\keyword{datasets}
