% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiny.lists.and.enums.R
\docType{data}
\name{MFMRDates}
\alias{MFMRDates}
\title{The Date-Time Formats for use in R (the \code{MFMRutils} selection)}
\format{
An object of class \code{environment} of length 3.
}
\usage{
MFMRDates
}
\description{
A collection of frequently used Date-Time (text) Formats to support the MFMR
Suite of R Functions.
}
\examples{
### Easily print & assign icons as follows ...
MFMRDates$BOLD      ### -> sets the text font format to BOLD !!!
MFMRDates$ITALICS   ### -> sets the text font format to ITALICS !!!
MFMRDates$RESET     ### -> removes actively applied ANSI Text Formatting !!!

### Use with the direct-access R operator "::" from anywhere ...
MFMRutils::MFMRFormat$BOLD   ### -> sets the text font format to BOLD !!!

}
\keyword{datasets}
