% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKGS.Check.Code.Specs.R
\name{pkgs.check.code.specs}
\alias{pkgs.check.code.specs}
\title{CRAN Code Validation with File Version Tracking ("SuiteMFMR" DevTools)}
\usage{
pkgs.check.code.specs(
  sbCheckDocs = TRUE,
  sbCheckCRAN = FALSE,
  ssTimeZone = "Africa/Windhoek",
  ...
)
}
\arguments{
\item{sbCheckDocs}{a logical (boolean) value that specifies whether to run the
standard package documentation process (only - not the FULL
CRAN Code Validation Process/Check).}

\item{sbCheckCRAN}{a logical (boolean) value that specifies whether to run the
FULL CRAN Code Validation Process/Check (inclusive of the
pre-check documentation steps).}

\item{ssTimeZone}{a simple character vector (string) that defines the Time Zone
to be used for the package documentation.}

\item{...}{the fall-through (DotsArgs) function arguments used to prime nested
functions within this main function. These "DotsArgs" pertain mostly
to the "Func-SELF-ID" SuiteMFMR helper function.}
}
\value{
\itemize{
\item This function returns the programmatically amended or updated (real-time or
active) version number for the active R Library Project as a list of character
objects.
\item This function creates a Work-In-Progress (WIP) directory at the root of the
active R-Libs Project (i.e. "./WIP" <- if not already exists). The "./WIP"
folder and all of its contents are included in GIT push-&-pull processes. You
must MANUALLY ADD the following code stub (^WIP$) \if{html}{\out{<sans parentheses>}} to the
".Rbuildignore" R Project file to ensure the "./WIP" folder + contents are
excluded from the R Project Build Process. (<- this is IMPORTANT !!!)
\item This function also creates a "DevsVersTimeStamp.txt" file in the "./WIP"
folder for secondary development version tracking purposes (as needed).
}
}
\description{
A Helper Function that executes the CRAN pre-requisite Code Checking Procedure
during active R Package Development. This function programmatically updates the
package version number in the R Project DESCRIPTION file before running the
required documentation and/or CRAN Package Pre-Submission Code Requirement
Checks (i.e. CRAN Code Validation) during iterative development cycles. This
Helper Function was originally crafted as technical support tool for the "MFMR
Suite of R Libraries" (aka "SuiteMFMR"), but may also be helpful as a support
tool during the R Package Development Cycle of other (3rd Party) R Projects.
}
\examples{
### Run R Package DevCode easily as follows ...
library(MFMRutils)   # <- Loads "MFMRutils" library (if already installed) !!!

### Run 2 different types of code check/validation processes ...
pkgs.check.code.specs(sbCheckDocs = TRUE)   # -> Executes only the DevTools Documentation Process.
pkgs.check.code.specs(sbCheckCRAN = TRUE)   # -> Executes the more complete CRAN Code Validation Process.

### Check (i.e. "rasDevToolsCHECK()") overrides the documentation process ...
# The Documentation Process will only be executed once if both are TRUE !!!
pkgs.check.code.specs(sbCheckDocs = TRUE, sbCheckCRAN = TRUE)

}
