% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKGS.Get.Lib.Info.R
\name{pkgs.get.lib.info}
\alias{pkgs.get.lib.info}
\title{Extract R Package \if{html}{\out{<Library>}} Information}
\usage{
pkgs.get.lib.info(ssPathFileDESC = "./DESCRIPTION", sbIsAbsolutePath = TRUE)
}
\arguments{
\item{ssPathFileDESC}{a character (string) value that denotes the \if{html}{\out{<local>}} path
\if{html}{\out{<directory>}} to the R Library "DESCRIPTION" file.}

\item{sbIsAbsolutePath}{a logical (boolean) value that specifies whether the
DESCRIPTION File Path argument \code{ssPathFileDESC} is set
(or provided) as an absolute path or not.}
}
\description{
A tiny R function that extracts relevant package information from \if{html}{\out{<internal>}}
R Library Attributes.
}
\examples{
### Activate the "MFMRutils" R Library (if previously installed) ...
library(MFMRutils)   # -> Loads the "MFMRutils" library !!!

### Extract the R Library Information (for active R Library <project>) ...
rlsLibINFO <- pkgs.get.lib.info()

rlsLibINFO[["NAME"]]      # <- Returns 'Package Name' value <result> ...
rlsLibINFO[["VERSION"]]   # <- Returns 'Package Version' value <result> ... 

}
