% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projs.check.libs.R
\name{projs.check.libs}
\alias{projs.check.libs}
\title{Install required R Project Libraries}
\usage{
projs.check.libs(
  vsReqLibs = NULL,
  sbFixLibs = FALSE,
  sbQuietInstall = FALSE,
  sbShowLibs = TRUE,
  ssFuncSelfID = "Check Proj. Libs",
  sbRunSelfID = FALSE,
  ssFuncCallerID = NULL,
  sbUpdateLibs = TRUE,
  ssFuncType = NULL,
  ssFormatDT = "\%a, \%b \%d \%Y \%X"
)
}
\arguments{
\item{vsReqLibs}{a vector argument containing the discrete (i.e. comma-separated)
character values (i.e. strings or texts) of the 3rd Party R
Libraries required for the active R project.}

\item{sbFixLibs}{a logical (boolean) argument that specifies whether to fix (
i.e. install or update) any missing required R libraries.}

\item{sbQuietInstall}{a logical (boolean) argument that specifies whether the
R library install or update should be verbose or not.}

\item{sbShowLibs}{a logical (boolean) argument specifies whether the input R
libraries (i.e. required R libraries) should be printed out
to the R Console or returned to the function output (result)
at the completion of the function code execution.}

\item{ssFuncSelfID}{a character vector (i.e. string or text) serving as a
self-identifier (tag) for this "projs.check.libs()" function.}

\item{sbRunSelfID}{a logical (boolean) argument that specifies whether the
"projs.check.libs()" function should output its own identifier
information alongside the user specified info (notification).}

\item{ssFuncCallerID}{a character vector (string or text) serving as the
identifier (tag) for the R Function or R Project that
called (invoked) this "projs.check.libs()" function.}

\item{sbUpdateLibs}{a logical (boolean) argument that specifies whether any
outdated R libraries should be updated or not. If set to
\link{FALSE} the function will merely post all outdated R libs
to the R Console and not take any further action towards
"fixing" the outdated R libraries.}

\item{ssFuncType}{a character vector (string or text) that specifies the
function type. OPTIONS: "Helper" (small to medium custom
functions; less than 300 lines of code) and "LARGE" (massive
custom function with >300 lines of code). If set to NULL
the function type will be assigned the default value of "Helper".}

\item{ssFormatDT}{a character vector (string or text) that specifies the
DateTime format to be used for displaying date-times in
the console.}
}
\value{
The function outputs a list containing to list items ...
\itemize{
\item ReqLibs: a vector containing the R library IDs supplied via the
"vsReqLibs" function argument.
\item FI: additional information on the functions internal code execution
process.
}
}
\description{
A Helper Function to install 3rd Party R Libraries required by an R Project
but that are not currently installed. This custom function was intended to
mainly support the MFMR Suite of R Functions, but can be utilized as a
standalone function in other R packages.
}
\examples{
### Easily install missing and update outdated R Project Libraries ...
library(MFMRutils)   # <= Load the `MFMRutils` library (if previously installed).


## Explicitly set a CRAN mirror ... (<- only needed for CRAN-CODE-CHECKS) !!!
options(repos = c(CRAN = "https://cloud.r-project.org"))


### Install the required R Project libraries as follows ...
vsReqLibs_ <- c("ggplot2", "dplyr")   # <- Specify the required R Libraries accordingly !!!
projs.check.libs(           # <= Run the function with the "vsReqLibs" function argument
  vsReqLibs = vsReqLibs_,   #    provided accordingly ...
  sbFixLibs = TRUE   # <= Set the "sbFixLibs" function argument to TRUE to ensure all
)                    #    missing R Libraries are installed !!!



### Update outdated R Project libraries as follows ...
vsReqLibs_ <- c("bench", "zoo")   # <- Specify the required R Libraries accordingly !!!
projs.check.libs(           # <= Run the function with the "vsReqLibs" function argument
  vsReqLibs = vsReqLibs_,   #    provided accordingly !!!
  sbFixLibs = TRUE,     # <= Set both the "sbFixLibs" and "sbUpdateLibs" function
  sbUpdateLibs = TRUE   #    arguments to TRUE to ensure all outdated R libraries
)

}
