% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROJS.Check.Proj.Libs.R
\name{projs.check.proj.libs}
\alias{projs.check.proj.libs}
\title{Install Required R Project Libraries}
\usage{
projs.check.proj.libs(
  vsReqLibs = NULL,
  sbFixLibs = FALSE,
  sbQuietInstall = FALSE,
  sbShowLibs = TRUE,
  ssFuncSelfID = "MFMRutils-Check.Libs",
  sbRunSelfID = FALSE,
  ssFuncCallerID = NULL,
  sbUpdateLibs = FALSE,
  ssFormatDT = "\%a, \%b \%d \%Y \%X",
  sbPrintPretty = TRUE,
  ...
)
}
\arguments{
\item{vsReqLibs}{a vector of strings (characters) containing the discrete (i.e.
comma-separated) character values of the R Libraries required.}

\item{sbFixLibs}{a logical (boolean) value that specifies whether to fix (i.e.
install or update) any missing R libraries as specified in
the 'vsReqLibs' function argument.}

\item{sbQuietInstall}{a logical (boolean) value that specifies whether the R
library install or update should be verbose \link{FALSE} or
not \link{TRUE}.}

\item{sbShowLibs}{a logical (boolean) value that specifies whether the input R
libraries (i.e. 'vsReqLibs') should be printed out to the R
Console or returned to the function output (result) upon
completion of the function code execution. If set to \link{TRUE}
then both actions (i.e. show & return) will be executed.}

\item{ssFuncSelfID}{a character vector (i.e. string or text) serving as a
self-identifier (tag) for this "projs.check.libs()" function.}

\item{sbRunSelfID}{a logical (boolean) argument that specifies whether the
"projs.check.libs()" function should output its own identifier
information alongside the user specified info (notification).}

\item{ssFuncCallerID}{a character vector (string or text) serving as the
identifier (tag) for the R Function or R Project that
called (invoked) this "projs.check.libs()" function.}

\item{sbUpdateLibs}{a logical (boolean) value that specifies whether any
outdated R libraries should be updated or not. If set to
\link{FALSE} the function will merely post all outdated R libs
to the R Console and not take any further action towards
"fixing" the outdated R libraries.}

\item{ssFormatDT}{a character vector (string or text) that specifies the
DateTime format to be used for displaying date-times in
the console.}

\item{sbPrintPretty}{a logical (boolean) argument that specifies whether the
ANSI text font formatting should be applied to the
printed notification or not.}
}
\value{
The function outputs a list containing to list items ...
\itemize{
\item ReqLibs: a vector containing the R library IDs supplied via the
"vsReqLibs" function argument.
\item FI: additional information on the functions internal code execution
process.
}
}
\description{
A Helper Function to install 3rd Party R Libraries required by an R Project
but that are not currently installed. This custom function was intended to
mainly support the MFMR Suite of R Functions, but can be utilized as a
standalone function in other R packages.
}
\examples{
### Easily install missing and update outdated R Project Libraries ...
library(MFMRutils)   # -> Load the `MFMRutils` library (if previously installed).


## Explicitly set a CRAN mirror ... (<- only needed for CRAN-CODE-CHECKS) !!!
options(repos = c(CRAN = "https://cloud.r-project.org"))


### Install the required R Project libraries as follows ...
vsReqLibs_ <- c("ggplot2", "dplyr")   # <- Specify the required R Libraries accordingly !!!
projs.check.libs(           # -> Run the function with the "vsReqLibs" function argument
  vsReqLibs = vsReqLibs_,   #    provided accordingly ...
  sbFixLibs = TRUE   # -> Set the "sbFixLibs" function argument to TRUE to ensure
)                    #    all missing R Libraries are installed !!!



### Update outdated R Project libraries as follows ...
vsReqLibs_ <- c("bench", "zoo")   # -> Specify the required R Libraries accordingly !!!
projs.check.libs(           # -> Run the function with the "vsReqLibs" function argument
  vsReqLibs = vsReqLibs_,   #    provided accordingly !!!
  sbFixLibs = TRUE,     # -> Set both the "sbFixLibs" and "sbUpdateLibs" function
  sbUpdateLibs = TRUE   #    arguments to TRUE to ensure all outdated R libraries
)

}
