% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.tiny.helpers.R
\name{"pkgs.get.lib.info"}
\alias{"pkgs.get.lib.info"}
\alias{pkgs.get.lib.info}
\title{Extract R Package \if{html}{\out{<Library>}} Information}
\usage{
pkgs.get.lib.info(ssPathFileDESC = "./DESCRIPTION")
}
\arguments{
\item{ssPathFileDESC}{a character (string) value that denotes the \if{html}{\out{<local>}} path
\if{html}{\out{<directory>}} to the R Library "DESCRIPTION" file.}
}
\description{
A tiny R function that extracts relevant package information from \if{html}{\out{<internal>}}
R Library Attributes.
}
\examples{
### Activate the "MFMRutils" R Library (if previously installed) ...
library(MFMRutils)   # -> Loads the "MFMRutils" library !!!

### Extract the R Library Information (for active R Library <project>) ...
rlsLibINFO <- pkgs.get.lib.info()

rlsLibINFO[["PkgNAME"]]      # <- Returns 'Package Name' value <result> ...
rlsLibINFO[["PkgVERSION"]]   # <- Returns 'Package Version' value <result> ...


}
