% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RENV_COLOURS.R
\docType{data}
\name{RENV_COLOURS}
\alias{RENV_COLOURS}
\title{ANSI Colours for use in R (the "SuiteMFMR" selection)}
\format{
An object of class \code{environment} of length 32.
}
\usage{
RENV_COLOURS
}
\value{
This R Object returns an R Environment-Locked List of ANSI Colour Codes that may be applied to
both foreground and background R Console \if{html}{\out{<visual>}} Layers and Texts.
}
\description{
A collection of frequently used ANSI (American National Standards Institute) Colours to support
the "MFMR Suite of R Functions" (a.k.a. "SuiteMFMR").
}
\section{CONSTANTS}{
NA
}

\examples{
### Load the required R Library ...
library(MFMRutils)   # <- Loads library (if already installed locally) !!!

### ... the easily assign COLORS as follows ...
RENV_COLOURS$BlackFORE     ### -> outputs the text in black font colour !!!
RENV_COLOURS$CyanFORE      ### -> outputs the text in cyan font colour !!!
RENV_COLOURS$MagentaBACK   ### -> outputs the text background in black font colour !!!

### Use with the direct-access R operator "::" from anywhere ...
MFMRutils::RENV_COLOURS$CyanFORE   ### -> outputs the text in cyan font colour !!!


}
\seealso{
Other SuiteMFMR CONSTANTS: 
\code{\link{RENV_DEVS}},
\code{\link{RENV_FORMATS}},
\code{\link{RENV_FSID}},
\code{\link{RENV_ICONS}},
\code{\link{RENV_SuiteMFMR}}
}
\concept{SuiteMFMR CONSTANTS}
\keyword{datasets}
