% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RENV_FORMATS.R
\docType{data}
\name{RENV_FORMATS}
\alias{RENV_FORMATS}
\title{Text & Date Formats for use in R (the "SuiteMFMR" selection)}
\format{
An object of class \code{environment} of length 11.
}
\usage{
RENV_FORMATS
}
\value{
This R Object returns an R Environment-Locked List of
}
\description{
A collection of frequently used ANSI (American National Standards Institute) Text Font and POSIX
Date Formats to support the "MFMR Suite of R Functions" (a.k.a. "SuiteMFMR").
}
\section{CONSTANTS}{
NA
}

\examples{
### Load the required R Library ...
library(MFMRutils)   # <- Loads library (if already installed locally) !!!

### ... then easily apply TEXT FORMATS as follows ...
RENV_FORMATS$ANSI_BOLD      # -> sets the text font format to ANSI_BOLD !!!
RENV_FORMATS$ANSI_ITALICS   # -> sets the text font format to ANSI_ITALICS !!!
RENV_FORMATS$ANSI_RESET     # -> removes actively applied ANSI Text Formatting !!!

### Use with the direct-access R operator "::" from anywhere ...
MFMRutils::RENV_FORMATS$ANSI_BOLD   # -> sets the text font format to ANSI_BOLD !!!


}
\seealso{
Other SuiteMFMR CONSTANTS: 
\code{\link{RENV_COLOURS}},
\code{\link{RENV_DEVS}},
\code{\link{RENV_FSID}},
\code{\link{RENV_ICONS}},
\code{\link{RENV_SuiteMFMR}}
}
\concept{SuiteMFMR CONSTANTS}
\keyword{datasets}
