% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RENV_FSID.R
\docType{data}
\name{RENV_FSID}
\alias{RENV_FSID}
\title{List of Function Self-ID Argument Tags ("SuiteMFMR" DevTools)}
\format{
An object of class \code{environment} of length 43.
}
\usage{
RENV_FSID
}
\value{
This R Object returns an R Environment-Locked List of
}
\description{
A collection of frequently used ANSI (American National Standards Institute)
Text Font Formats to support the "MFMR Suite of R Functions" (aka "SuiteMFMR").
}
\section{CONSTANTS}{
NA
}

\examples{
### Install the required R Library ...
require(MFMRutils)   # <- Ensures the "MFMRutils" library is installed locally !!!

### Use with the direct-access R operator "::" from anywhere ...
RENV_FSID$PROJ_ID     ### -> sets the text font format to BOLD !!!

### Use with the direct-access R operator "::" from anywhere ...
MFMRutils::RENV_FSID$PROJ_ID     ### -> sets the text font format to BOLD !!!
MFMRutils::RENV_FSID$PROJ_ID     ### -> sets the text font format to BOLD !!!
MFMRutils::RENV_FSID$FUNC_MODE   ### -> sets the text font format to ITALICS !!!
MFMRutils::RENV_FSID$TIME_STOP   ### -> removes actively applied ANSI Text Formatting !!!


}
\seealso{
Other SuiteMFMR CONSTANTS: 
\code{\link{RENV_COLOURS}},
\code{\link{RENV_DEVS}},
\code{\link{RENV_FORMATS}},
\code{\link{RENV_ICONS}},
\code{\link{RENV_SuiteMFMR}}
}
\concept{SuiteMFMR CONSTANTS}
\keyword{datasets}
