% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RENV_ICONS.R
\docType{data}
\name{RENV_ICONS}
\alias{RENV_ICONS}
\title{Icons For Use in R (the "SuiteMFMR" selection)}
\format{
An object of class \code{environment} of length 33.
}
\usage{
RENV_ICONS
}
\value{
This R Object returns an R Environment-Locked List of Text Icons, in unicode form, used widely
throughout the "MFMR Suite of R Packages" (a.k.a. "SuiteMFMR").
}
\description{
A collection of text icons (in unicode format) meant to support the "MFMR Suite of R Functions"
(a.k.a. "SuiteMFMR").
}
\section{CONSTANTS}{
NA
}

\examples{
### Load the required R Library ...
library(MFMRutils)   # <- Loads library (if already installed locally) !!!

### ... then easily assign ICONS as follows ...
RENV_ICONS$OoglyEyes      # -> prints "👀" to the R console !!!
RENV_ICONS$CheckMark      # -> prints "✔" to the R console !!!
RENV_ICONS$SmileyPonder   # -> prints "🤔" to the R console !!!

### Use with the direct-access R operator "::" from anywhere ...
MFMRutils::RENV_ICONS$OoglyEyes   ### -> prints "👀" to the R console !!!


}
\seealso{
Other SuiteMFMR CONSTANTS: 
\code{\link{RENV_COLOURS}},
\code{\link{RENV_FORMATS}},
\code{\link{RENV_FSID}},
\code{\link{RENV_MISC}},
\code{\link{RENV_SuiteMFMR}}
}
\concept{SuiteMFMR CONSTANTS}
\keyword{datasets}
