% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RENV_MISC.R
\docType{data}
\name{RENV_MISC}
\alias{RENV_MISC}
\title{The "SuiteMFMR" List of Useful Constants}
\format{
An object of class \code{environment} of length 12.
}
\usage{
RENV_MISC
}
\value{
This R Object returns an R Environment-Locked List of CONSTANTS (encapsulating character, numeric
and other R Objects \if{html}{\out{<values>}}) used extensively throughout the "SuiteMFMR" R Project.
}
\description{
A collection of useful CONSTANTS that are utilized throughout the "MFMR Suite of R Packages"
(i.e. "SuiteMFMR"). This Environment-LOCKED R List standardizes and improves the use of key
values required by multiple "SuiteMFMR" functions.
}
\section{CONSTANTS}{
NA
}

\examples{
### Install the required R Library ...
require(MFMRutils)   # <- Ensures the "MFMRutils" library is installed locally !!!

### Use with the direct-access R operator "::" from anywhere ...
RENV_MISC$PATH_FOLDER_WIP_PROD

### ... then easily apply TEXT FORMATS as follows ...
RENV_MISC$PATH_FOLDER_WIP_PROD   # -> Returns the Work-In-Progress Project PATH (i.e. "./WIP) 
                                   #    for the R Library Code Development Staging Folder (i.e. 
                                   #    "./WIP/01_NextForPROD") !!!


}
\seealso{
Other SuiteMFMR CONSTANTS: 
\code{\link{RENV_COLOURS}},
\code{\link{RENV_FORMATS}},
\code{\link{RENV_FSID}},
\code{\link{RENV_ICONS}},
\code{\link{RENV_SuiteMFMR}}
}
\concept{SuiteMFMR CONSTANTS}
\keyword{datasets}
