% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RENV_SuiteMFMR.R
\docType{data}
\name{RENV_SuiteMFMR}
\alias{RENV_SuiteMFMR}
\title{Full List of "SuiteMFMR" R Project Packages}
\format{
An object of class \code{environment} of length 7.
}
\usage{
RENV_SuiteMFMR
}
\value{
This R Object returns an R Environment-Locked List of lists containing the names and URLs of all
the R Libraries that form part of the "SuiteMFMR" R Project.
}
\description{
The complete list of R Libraries that constitute the "SuiteMFMR" R Project. This Environment-
Locked List contains the "standard" IDs (i.e. "tags" or names) and URLs (i.e. the online source
repository links) for the "MFMR Suite of R Packages" (a.k.a. "SuiteMFMR").
}
\section{CONSTANTS}{
NA
}

\examples{
### Load the required R Library ...
library(MFMRutils)   # <- Loads library (if already installed locally) !!!

### ... then easily extract the "SuiteMFMR" R Library IDs as follows ...
RENV_SuiteMFMR$UTILS   # -> extracts the "Utils" MFMR R Library Name and URL information !!!
RENV_SuiteMFMR$PLOTS   # -> extracts the "Plots" MFMR R Library Name and URL information !!!
RENV_SuiteMFMR$STATS   # -> extracts the "Stats" MFMR R Library Name and URL information !!!


}
\seealso{
Other SuiteMFMR CONSTANTS: 
\code{\link{RENV_COLOURS}},
\code{\link{RENV_FORMATS}},
\code{\link{RENV_FSID}},
\code{\link{RENV_ICONS}},
\code{\link{RENV_MISC}}
}
\concept{SuiteMFMR CONSTANTS}
\keyword{datasets}
