% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiny.lists.and.enums.R
\docType{data}
\name{ANSIsColors}
\alias{ANSIsColors}
\title{ANSI Colours for use in R ...}
\format{
An object of class \code{environment} of length 32.
}
\usage{
ANSIsColors   ### <- if "MFMRutils" library is already installed & loaded !!!
}
\description{
A collection of frequently used ANSI (American National Standards Institute) Colours to support the MFMR Suite of R Functions.
}
\examples{
### Easily print & assign icons as follows ...
ANSIsColors$BlackFORE     ### <- colours the text in black font colour !!!
ANSIsColors$CyanFORE      ### <- colours the text in cyan font colour !!!!!!
ANSIsColors$MagentaBACK   ### <- colours the text background in black font colour !!! !!!

### Use with the direct-access R operator "::" from anywhere ...
MFMRutils::ANSIsColors$OoglyEyes   ### <- prints "👀" to the R console !!!

}
\keyword{datasets}
