% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiny.lists.and.enums.R
\docType{data}
\name{ANSIsFormat}
\alias{ANSIsFormat}
\title{ANSI Text Formats for use in R ...}
\format{
An object of class \code{environment} of length 5.
}
\usage{
ANSIsFormat   ### <- if "MFMRutils" library is already installed & loaded !!!
}
\description{
A collection of frequently used ANSI (American National Standards Institute) Text Font Formats to support the MFMR Suite of R Functions.
}
\examples{
### Easily print & assign icons as follows ...
ANSIsFormat$BOLD      ### <- set the text font format to BOLD !!!
ANSIsFormat$ITALICS   ### <- set the text font format to ITALICS !!!
ANSIsFormat$RESET     ### <- removes any actively applied ANSI Text Formatting !!!

### Use with the direct-access R operator "::" from anywhere ...
MFMRutils::ANSIsFormat$BOLD   ### <- set the text font format to BOLD !!!

}
\keyword{datasets}
