% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CODE.Clean.File.Path.R
\name{code.clean.file.path}
\alias{code.clean.file.path}
\title{Sanitize String and Vector File Paths}
\usage{
code.clean.file.path(ssPathString = NULL, vsPathVector = NULL)
}
\arguments{
\item{ssPathString}{a dynamic (complex) object that captures the "Left-Hand" R Object
to be evaluated against the "NULL Criteria" (i.e. by way of value,
length or class \if{html}{\out{<data type>}} NULL Checks).}

\item{vsPathVector}{a dynamic (complex) object that captures the "Right-Hand" R Object
(to be returned if the "Left-Hand" R Object is NULL).}
}
\description{
The "SuiteMFMR" Null-Coalescing Operator (NCO) is similar to the "??" NCO of
the DART Programming Language. The operator evaluates whether the "Left-Hand" R
Object is NULL (in terms of its value or length or class \if{html}{\out{<data type>}}) and returns
the "Right-Hand" R Object if the "Left-Hand" Object is \if{html}{\out{<indeed>}} NULL.
}
\examples{
### Use the Null-Coalescing Operator (NCO) as follows: ...
library(MFMRutils)   # <- Loads the "MFMRutils" library (if already installed) ...

### Then apply the NCO accordingly ...
NULL \%??\% "DeFAULT"                    # -> returns "Default" !!!
"ACTual" \%??\% "DEFault"                # -> returns "ACTual" !!!
NULL \%??\% NULL \%??\% "FINal"            # -> returns "FINal" !!!
NULL \%??\% "PENULTimate" \%??\% "FINal"   # -> returns "PENULTimate" !!!

}
