% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CODE.Get.CELN.R
\name{code.get.celn}
\alias{code.get.celn}
\title{Extract the Code Editor Gutter Line Number (CELN)}
\usage{
code.get.celn(ssFuncName = "code.get.celn", siStartCELN = 1L, siCallIndex = 1L)
}
\arguments{
\item{ssFuncName}{(\link{character}) A String value that specifies the Function Identifier (i.e. name)
where the R Code Search should be conducted (directed at).}

\item{siStartCELN}{(\link{integer}) An Integer value that specifies the Code Editor Gutter Line Number
(abbreviated as \code{CELN}) where the function's opening (i.e. starting) curly
brace is located (i.e. the code editor line number at which the function
body's block of code starts !!!).}

\item{siCallIndex}{(\link{integer}) An Integer value that defines the "Call Index" in the localized (
function \if{html}{\out{<internal>}}) call stack in reference to when (i.e. in which order) the
\code{code.get.celn()} function was called (i.e. from inside a function).}
}
\description{
A SPECIAL HELPER Function that extracts the Code Editor Gutter Line Number (CEGLN a.k.a. CELN) at
the code editor line location from which the \code{code.get.celn()} function was called. This function
was created for the purpose of enhancing the debugging, R Package development and real-time (i.e.
interactive) code documentation processes from within the MFMR Suite of R Functions. However,
this function may also be used as a stand-alone function by (or in) other, 3rd Party, libraries
and custom R functions - as long as the \code{MFMRutils} library is linked (i.e. previously installed
and/or specified as a library dependency).
}
\examples{
### Easily debug custom R function code with this <cool> helper function ...
library(MFMRutils)   # <- Ensures the `MFMRutils` library is <already> installed locally.



### Use with any custom R function as follows ...
"my.cust.r.func" <- function(x=7, y=3, z=28) {   # <- Use the Code Editor Line Number (CELN)
                                                 #    at which the `function()` call's opening 
                                                 #    brace or bracket <i.e. `(`> is located at 
                                                 #    as the `siStartCELN` input argument value.
  siStartCELN <- 19L; # <- 👆 assumes this opening brace above (denoting the start of
                      #    the <custom> function's arguments code block) is located
                      #    at line 19 of the code editor (i.e. the opening brace is
                      #    located at the 19th CELN).

  RCT_TAG_FUNC_ID <- "My.Cust.R.FUNC";  # <- ALWAYS TAG (ID) Custom R Functions accordingly !!!

  valSUM <- sum(x, y, z);
  cat(
    paste0(
      " \u279C ", RCT_TAG_FUNC_ID, " " , 
      MFMRutils::code.get.celn(
        RCT_TAG_FUNC_ID,   # <- Provide the Function ID (tag) for the Custom R Function ...
        siStartCELN,       # <- Specify at which line the opening function brace is located !!! 
        1L                 # <- Define the call order <call sequence> for the `CODE.Get.CELN()` 
      ),                   #    function calls inside the custom function code. This is the 1st 
                           #    time the `CODE.Get.CELN()` function is called inside this custom 
                           #    function so the `siCallIndex` value here should == `1L`.
      " | Summed all 3 input values < result: ", valSUM," > !!! \n"
    )
  );

  valMEAN <- sum(x, y, z) / 3;
  cat(
    paste0(
      " \u279C ", RCT_TAG_FUNC_ID, " " ,
      MFMRutils::code.get.celn(
        RCT_TAG_FUNC_ID,   # <- Provide the Function ID (tag) for the Custom R Function ...
        siStartCELN,       # <- Specify at which line the opening function brace is located !!! 
        2L                 # <- Define the call order <call sequence> for the `CODE.Get.CELN()` 
      ),                   #    function calls inside the custom function code. This is the 2nd 
                           #    time the `CODE.Get.CELN()` function is called inside this custom 
                           #    function so the `siCallIndex` value here should == `2L`.
      " | Averaged the 3 input values < result: ", round(valMEAN, 1)," > !!! \n"
    )
  );
  
  # Notify function's SUCCESSFUL COMPLETION ( <- is only needed for code illustration) ...
  cat(
    paste0(
      " \u279C ", RCT_TAG_FUNC_ID, " " ,
      MFMRutils::code.get.celn(
        RCT_TAG_FUNC_ID,   # <- Provide the Function ID (tag) for the Custom R Function ...
        siStartCELN,       # <- Specify at which line the opening function brace is located !!! 
        3L                 # <- Define the call order <call sequence> for the `CODE.Get.CELN()` 
      ),                   #    function calls inside the custom function code. This is the 3rd 
                           #    time the `CODE.Get.CELN()` function is called inside this custom 
                           #    function so the `siCallIndex` value here should == `3L`.
      " | CUSTOM R-FUNC `", tolower(RCT_TAG_FUNC_ID), "()` SUCESSFULLY EXECUTED !!! \n\n"
    )
  );

  return(
    list("SUM" = valSUM, "MEAN" = valMEAN)
  );
}

## Execute the custom R Function ...
my.cust.r.func()

## Outputs from "my.cust.r.func()" ...
# ➜ My.Cust.R.FUNC 34 | Summed all 3 input values < result: 38 > !!!
# ➜ My.Cust.R.FUNC 49 | Averaged the 3 input values < result: 12.7 > !!!
# ➜ My.Cust.R.FUNC 64 | CUSTOM R-FUNC `my.cust.r.func()` SUCESSFULLY EXECUTED !!!

# $SUM
# [1] 45

# $MEAN
# [1] 15


}
\seealso{
Other SuiteMFMR Code Functions: 
\code{\link{\%??\%}()},
\code{\link{code.append.text.to.file}()},
\code{\link{code.classify.func}()},
\code{\link{code.clean.file.path}()},
\code{\link{code.return.renv.locked.list}()},
\code{\link{code.split.string.on.vector}()}
}
\concept{SuiteMFMR Code Functions}
