% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEVS.Patch.Code.Dev.Trckr.File.R
\name{devs.patch.code.dev.trckr.file}
\alias{devs.patch.code.dev.trckr.file}
\title{R Code Development Tracker File Updater ("SuiteMFMR" DevTools)}
\usage{
devs.patch.code.dev.trckr.file(sbIsProdRel = FALSE, sbAudioNote = FALSE)
}
\arguments{
\item{sbIsProdRel}{a logical (boolean) value that captures if the code-check and/or code commit
process (action) is a "Production Release" action or not.}

\item{sbAudioNote}{a logical (boolean) value that specifies whether an audio notification should
be played at the completion of this function (i.e. upon successful patching or
updating of the "Active Development Information Tracker File").}
}
\value{
\itemize{
\item This function programmatically amends \if{html}{\out{<patches or updates>}} the active (i.e. regent) R Library
Project's Code Development Tracking Information (i.e. tracker file) for code versioning and
project (i.e. library code) development tracking purposes.
}
}
\description{
A Helper Function that updates important information inside the "Active Development Information
Tracker File" (i.e. "00_ACT_DEV_TRCKR.txt") located in the active project's development or
launchpad folder (i.e. the "./WIP/" path or directory). This "Active Development Information
Tracker File" will be programmatically initialized (created) by this helper function if not
found in the default project directory (i.e. "./WIP/" project folder).
}
\examples{
### Run R Package DevCode easily as follows ...
library(MFMRutils)   # <- Loads "MFMRutils" library (if already installed) !!!

### Run 2 different types of code check/validation processes ...
devs.patch.code.dev.trckr.file()   # -> Executes only the DevTools Documentation 
                                   #    Process.

}
