% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEVS.Pull.Libr.Info.R
\name{devs.pull.libr.info}
\alias{devs.pull.libr.info}
\title{Extract R Package Meta-Information ("SuiteMFMR" DevTools)}
\usage{
devs.pull.libr.info(ssPathFileDESC = "DESCRIPTION", sbIsAbsolutePath = FALSE)
}
\arguments{
\item{ssPathFileDESC}{(\link{character}) A String (i.e. character vector) value that denotes the local
path \if{html}{\out{<directory>}} to the R Library "DESCRIPTION" file.}

\item{sbIsAbsolutePath}{(\link{logical}) A boolean value that specifies whether the DESCRIPTION File
Path argument \code{ssPathFileDESC} is set (or provided) as an absolute path
or not.}
}
\description{
A tiny R function that extracts relevant package information from the R Library Manifest or
R-Project Attributes File (i.e. the R Library DESCRIPTION File).
}
\examples{
\dontrun{   ### <- This function constitutes a development utility !!! This function requires a 
            ###    special development directory ("./WIP") that is created during the init-run
            ###    (initial R project setup) phase and is intended to facilitate a user-friendly
            ###    R Library development process. For these reasons the code examples below
            ###    should not be executed during "R_CMD_CHECK" code check procedures.
            
### Activate the "MFMRutils" R Library (if previously installed) ...
library(MFMRutils)   ### -> Loads the "MFMRutils" library !!!

### Extract the R Library Information (for active R Library <project>) ...
rlsLibINFO <- devs.pull.libr.info()  ### <- Extracts Library Information ... 

rlsLibINFO[["NAME"]]      # -> Returns 'Package Name' value <result> ...
rlsLibINFO[["VERSION"]]   # -> Returns 'Package Version' value <result> ... 

}

}
\seealso{
Other SuiteMFMR DevTools: 
\code{\link{devs.check.code.specs}()},
\code{\link{devs.create.r.function}()},
\code{\link{devs.find.code.instances}()},
\code{\link{devs.patch.code.dev.trckr.file}()}
}
\concept{SuiteMFMR DevTools}
