% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvDATEs.R
\docType{data}
\name{EnvDATES}
\alias{EnvDATES}
\title{The Date-Time Formats for use in R (the \code{MFMRutils} selection)}
\format{
An object of class \code{environment} of length 3.
}
\usage{
EnvDATES
}
\description{
A collection of frequently used Date-Time (text) Formats to support the MFMR
Suite of R Functions.
}
\examples{
### Easily print & assign Date Formats as follows ...
EnvDATES$LONGv01   ### -> formats Date to "Saturday, January 16 1982" !!!
EnvDATES$LONGv02   ### -> formats Date to "Sat, Jan 16 1982" !!!
EnvDATES$LONGv03   ### -> formats Date to "Sat, 16 Jan 1982" !!!

### Use with the direct-access R operator "::" from anywhere ...
MFMRutils::EnvDATES$LONGv02   ### -> formats Date to "Sat, Jan 16 1982" !!!

}
\keyword{datasets}
