% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvFORMATs.R
\docType{data}
\name{EnvFORMATS}
\alias{EnvFORMATS}
\title{ANSI Text Formats for use in R (the \code{MFMRutils} selection)}
\format{
An object of class \code{environment} of length 5.
}
\usage{
EnvFORMATS   ### -> if "MFMRutils" is already installed & loaded !!!
}
\description{
A collection of frequently used ANSI (American National Standards Institute)
Text Font Formats to support the MFMR Suite of R Functions.
}
\examples{
### Easily print & assign icons as follows ...
EnvFORMATS$BOLD      ### -> sets the text font format to BOLD !!!
EnvFORMATS$ITALICS   ### -> sets the text font format to ITALICS !!!
EnvFORMATS$RESET     ### -> removes actively applied ANSI Text Formatting !!!

### Use with the direct-access R operator "::" from anywhere ...
MFMRutils::EnvFORMATS$BOLD   ### -> sets the text font format to BOLD !!!

}
\keyword{datasets}
