% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENVFormats.R
\docType{data}
\name{ENVFormats}
\alias{ENVFormats}
\title{ANSI Text Formats for use in R (the \code{MFMRutils} selection)}
\format{
An object of class \code{environment} of length 5.
}
\usage{
ENVFormats   ### -> if "MFMRutils" is already installed & loaded !!!
}
\description{
A collection of frequently used ANSI (American National Standards Institute)
Text Font Formats to support the MFMR Suite of R Functions.
}
\examples{
### Easily print & assign icons as follows ...
ENVFormats$BOLD      ### -> sets the text font format to BOLD !!!
ENVFormats$ITALICS   ### -> sets the text font format to ITALICS !!!
ENVFormats$RESET     ### -> removes actively applied ANSI Text Formatting !!!

### Use with the direct-access R operator "::" from anywhere ...
MFMRutils::ENVFormats$BOLD   ### -> sets the text font format to BOLD !!!

}
\keyword{datasets}
