% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvICONs.R
\docType{data}
\name{EnvICONS}
\alias{EnvICONS}
\title{Unicode Icons for use in R (the \code{MFMRutils} selection)}
\format{
An object of class \code{environment} of length 9.
}
\usage{
EnvICONS   ### -> if "MFMRutils" library is already installed & loaded !!!
}
\description{
A collection of frequently used icons (Unicode characters) meant to support
the MFMR Suite of R Functions.
}
\examples{
### Easily print & assign icons as follows ...
EnvICONS$OoglyEyes      ### -> prints "ğŸ‘€" to the R console !!!
EnvICONS$CheckMark      ### -> prints "âœ”" to the R console !!!
EnvICONS$SmileyPonder   ### -> prints "ğŸ¤”" to the R console !!!

### Use with the direct-access R operator "::" from anywhere ...
MFMRutils::EnvICONS$OoglyEyes   ### -> prints "ğŸ‘€" to the R console !!!

}
\keyword{datasets}
