% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvLibsMFMR.R
\docType{data}
\name{EnvSuiteMFMR}
\alias{EnvSuiteMFMR}
\title{The MFMR Suite \if{html}{\out{<list>}} of R Libraries ...}
\format{
An object of class \code{environment} of length 7.
}
\usage{
EnvSuiteMFMR
}
\description{
The complete list of R Libraries that make up the "MFMR R Library Suite". This
Environment-Locked List contains the "standard" IDs for the MFMR R Libraries.
}
\examples{
### Load the required R Library ...
library(MFMRutils)   # <- Loads library (if already installed locally) !!!

### Easily extract the 'SuiteMFMR' R Library Identifiers <tags> as follows ...
EnvSuiteMFMR$UTILS   # -> extracts the 'Utils' MFMR R Library Identifier !!!
EnvSuiteMFMR$PLOTS   # -> extracts the 'Plots' MFMR R Library Identifier !!!
EnvSuiteMFMR$STATS   # -> extracts the 'Stats' MFMR R Library Identifier !!!

}
\keyword{datasets}
