% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INFO.Post.Func.Self.ID.R
\name{info.post.func.self.id}
\alias{info.post.func.self.id}
\title{Print Self-ID Information for Custom R Functions}
\usage{
info.post.func.self.id(
  ssProjID = NULL,
  ssFuncSelfID = NULL,
  siFuncMode01 = NULL,
  ssFuncCallerID = NULL,
  csIconCarat = NULL,
  csColorCarat = NULL,
  csIconSplit = NULL,
  csColorSplit = NULL,
  csTimeStart = NULL,
  csTimeStop = NULL,
  csFormatDT = NULL,
  csColorTimeStamp = NULL,
  csColorProjID = NULL,
  csColorFuncType = NULL,
  csColorCallerID = NULL,
  siStopCELN = NULL,
  siStartCELN = NULL,
  csColorMainText = NULL,
  sbPrintPretty = NULL,
  sbRunSelfID = FALSE
)
}
\arguments{
\item{ssProjID}{a character (string) identifier of the R Project \if{html}{\out{<script>}}
file that called the custom R function execution.}

\item{ssFuncSelfID}{a character (text string) identifier for the active R
Function being called (or being executed). If set to NULL
the hard-coded "info.post.func.self.id()" function ID
(i.e. "MFMR-Self.ID") will be used in the printed result.}

\item{siFuncMode01}{an integer (long) value that identifies the two run-time
states of this "Self-ID" custom function. A value of "1L"
sets the function mode to the "ENTER" (start) state; while
a value of "0L" sets the function mode to the "EXIT" (stop)
run-time state.}

\item{ssFuncCallerID}{a character (string) value that identifies the parent
function that invoked that execution of the active or
current function.}

\item{csIconCarat}{a character (string) value that defines the leading icon
(carat) object to be used when printing the "Self-ID"
information to the R Console.}

\item{csColorCarat}{a character (string) value that sets the "Carat Icon" object
colour in the formatted text output.}

\item{csIconSplit}{a character (string) value that defines the middle \if{html}{\out{<split>}}
icon (object) that should be used for printing the "Self-ID"
information to the R Console.}

\item{csColorSplit}{a character (string) value that sets the "Split" information
text colour in the formatted text output.}

\item{csTimeStart}{a character (string) value that defines the "START-Time"
of the active function.}

\item{csTimeStop}{a character (string) value that defines the "STOP-Time"
of the active function.}

\item{csFormatDT}{a character (string) value that specifies how "DateTime" info
should be formatted when printed out to the R Console.}

\item{csColorTimeStamp}{a character (string) value that sets the "TimeStamp"
information text colour in the formatted text output.}

\item{csColorProjID}{a character (string) value that sets the "ProjID" text
colour in the formatted text output.}

\item{csColorFuncType}{a character (string) value that sets the "FuncType" text
colour in the formatted text output.}

\item{csColorCallerID}{a character (string) value that sets the "CallerID" text
colour in the formatted text output.}

\item{siStopCELN}{an integer that denotes the STOP (end) of the function's
block of code. The Code Editor Line Number ("CELN") of the
last line of code that defines a custom function - usually
identified by a closing curly brace.}

\item{siStartCELN}{an integer that denotes the START of the function's block of
code. The Code Editor Line Number ("CELN") of the first line
of code that defines a custom function.}

\item{csColorMainText}{a character (string) value that sets the "Main" information
text colour in the formatted text output.}

\item{sbPrintPretty}{a logical (boolean) value that defines whether the output
text (Self-ID Function Info) should be formatted when
printed to the R Console.}

\item{sbRunSelfID}{a logical (boolean) value that defines whether the SELF-ID
procedure should be executed (TRUE) or not (FALSE).}
}
\value{
This function returns a single \if{html}{\out{<abbreviated>}} character value that classifies five function size
(code base size) classes for custom R functions: ...
\itemize{
\item "TNY" -> a "Tiny" R Function (less than 50 lines of code);
}
}
\description{
A simple Helper Function that compiles and prints the self-identification (i.e. self-id, type,
caller, run-time duration etc.) information of a custom R function. This custom function was
intended to mainly support the MFMR Suite of R Functions, but can be utilized as a standalone
function in other (i.e. 3rd Party) R packages.
}
\examples{
### Run Self-Identification (Self-ID) on your Custom R Function as follows:
library(MFMRutils)         # -> Loads the "MFMRutils" R Library ...
info.post.func.self.id()   # -> Runs the default <NULL> function state ...

### Prime the relevant function arguments as needed ...
info.post.func.self.id(
 ssProjID = "rTestProject",
 ssFuncSelfID = "rcfTextFUNC", 
 siFuncMode01 = 1L
)


}
\seealso{
Other SuiteMFMR INFO Functions: 
\code{\link{info.post.note}()}
}
\concept{SuiteMFMR INFO Functions}
