% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.post.note.R
\name{info.post.note}
\alias{info.post.note}
\title{A Helper Function that standardizes the User / Project Information Posting (i.e. notification) Processes.}
\usage{
info.post.note(
  ssPostNote = "NOTE to POST !!!",
  ssFuncSelfID = "Post Note",
  sbRunSelfID = FALSE,
  ssFuncCallerID = NULL,
  ssFuncType = NULL,
  siPostMode123 = 2L,
  sbRetFuncInfo = FALSE,
  ssPreStub = " => ",
  ssMidStub = " | ",
  sbPrePendNewLine = FALSE,
  sbPostPendNewLine = TRUE,
  sbEndExtraNewLine = FALSE
)
}
\arguments{
\item{ssPostNote}{the character vector (i.e. string or text or note) to be printed to the console. This text note is also returned as a function output under the "value" option of this function's results list.}

\item{ssFuncSelfID}{a character vector (i.e. string or text) serving as a self-identifier (tag) for this "info.post.note()" function.}

\item{sbRunSelfID}{a logical (i.e. boolean) value that specifies whether the "info.post.note()" function should output its own identifier information alongside the user specified info (notification).}

\item{ssFuncCallerID}{a character vector (string or text) serving as the identifier (tag) for the R Function or R Project that called (invoked) this "info.post.note()" function.}

\item{ssFuncType}{a character vector (string or text) that specifies the function type. OPTIONS: "Helper" (small to medium custom functions; less than 300 lines of code) and "LARGE" (massive custom function with >300 lines of code). If set to NULL the function type will be assigned the default value of "Helper".}

\item{siPostMode123}{an integer value that defines the form of the output (notification) message to be printed to the console. OPTIONS: 1L (applies the "START" or "HEADER" form to the printed message), 2L (applies the "Normal" or "DEFAULT" form to the printed message) and 3L (applies the "STOP" or "TERMINAL" form to the printed message). This function argument is extremely useful for defining the beginning and terminal sections of custom R Functions and/or R Projects.}

\item{sbRetFuncInfo}{a logical (boolean) value that specifies whether to output or return the function (i.e. internally processed) information in the results of the current function. This function argument can be useful when nesting multiple custom functions.}

\item{ssPreStub}{a character vector (string or text) that defines the starting text of the output message (notification). This function argument is useful in standardizing the notification format for an entire R Project.}

\item{ssMidStub}{a character vector (string or text) that defines the middle (ID "tag" and Messages) text of the output message (notification). This function argument is useful in standardizing the notification format for an entire R Project.}

\item{sbPrePendNewLine}{a logical (boolean) value that specifies whether a new line (blank space or row) should be added to the START of a posted note.}

\item{sbPostPendNewLine}{a logical (boolean) value that specifies whether a new line (blank space or row) should be added to the END of a posted note.}

\item{sbEndExtraNewLine}{a logical (boolean) value that specifies whether a second line (blank space or row) should be added to the END of a posted note. This function argument can be useful for clearly delineating sections within a custom R Function and/or R Project.}
}
\value{
\itemize{
\item This function prints the specified text (notification) directly to the console even if the function outputs are assigned to a variable.
\item This function also outputs additional function information (i.e. internally computed function information) as a list object.
}
}
\description{
A Helper Function that standardizes the User / Project Information Posting (i.e. notification) Processes.
}
\examples{
### Print a dummy notification ...
info.post.note()              # <= when "MFMRutils" library is loaded ...
MFMRutils::info.post.note()   # <= when "MFMRutils" library is NOT loaded !!!

### Print 3 different types of notifications ...
info.post.note(siPostMode123 = 1L)   # <= Prints a START (header) or beginning notification ...
info.post.note(siPostMode123 = 2L)   # <= Prints a Normal (default) or body notification ...
info.post.note(siPostMode123 = 3L)   # <= Prints an STOP (footer) or terminal notification ...

### Print additional (function internal) information ...
info.post.note(sbRetFuncInfo = T)   # <= Outputs additional function information in list form ...
}
