% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INFO.Post.Note.R
\name{info.post.note}
\alias{info.post.note}
\title{Post Standardized R-Project Notifications}
\usage{
info.post.note(
  csCarat = "=>",
  ssHead = NULL,
  csSplit = "|",
  ssBody = "NOTE to POST !!!",
  csTail = MFMRutils::RENV_ICONS$FireFlame,
  sbShowTail = TRUE,
  sbPrintPretty = TRUE,
  csColorHead = MFMRutils::RENV_COLOURS$GreenFORE,
  csColorBody = MFMRutils::RENV_COLOURS$CyanFORE,
  csColorCarat = MFMRutils::RENV_COLOURS$YellowFORE,
  csColorSplit = MFMRutils::RENV_COLOURS$YellowFORE,
  sbPrePendNL = FALSE,
  sbPostPend1NL = TRUE,
  sbPostPend2NLs = FALSE,
  sbPostAlways = TRUE,
  siCallCELN = 1982L
)
}
\arguments{
\item{csCarat}{(\link{character}) A complex String or Text (i.e. \link{character} \link{vector}) object that
defines what the first or starting symbol (i.e. carat) of the note, to be posted,
should be or look like. This can be a simple text or a complex (i.e. ANSI, RegEx
or escaped character) object - but must be an acceptable R \link{character} object.}

\item{ssHead}{(\link{character}) A simple String or Text (i.e. \link{character} \link{vector}) argument that
defines the starting (i.e. header) text of the note (to be posted). This function
argument is useful in standardizing the notification format at an R Project level.}

\item{csSplit}{(\link{character}) A complex String or Text (i.e. \link{character} \link{vector}) object that
defines what the middle symbol (i.e. separator between the \code{ssHead} and \code{ssBody}
components) of the note, to be posted, should be or look like. This can be a
simple text or a complex (i.e. ANSI, RegEx or escaped character) object - but must
be an acceptable R \link{character} object.}

\item{ssBody}{(\link{character}) A simple String or Text (i.e. \link{character} \link{vector}) argument that
defines the main (i.e. body) text of the note (to be posted).}

\item{csTail}{(\link{character}) A complex String or Text (i.e. \link{character} \link{vector}) object that
defines what the ending symbol (i.e. tail or terminal icon) of the note, to be
posted, should be or look like. This can be a simple text or a complex (i.e. ANSI,
RegEx or escaped character) object, but must be an acceptable R \link{character} object.}

\item{sbShowTail}{(\link{logical}) A Boolean value that specifies whether to include the tail (i.e.
trailing or terminal symbol or icon) object in the note to be posted or not
(default == \code{TRUE}).}

\item{sbPrintPretty}{(\link{logical}) A Boolean value that specifies whether the built-in (i.e. the
function's internal) text formatting (i.e. ANSI \if{html}{\out{<font weight and colour>}}
formats) should be applied to the printed notification or not (default ==
\code{TRUE}).}

\item{csColorHead}{(\link{character}) A String or Text (i.e. \link{character} \link{vector}) value that defines
the text colour for the header (\code{ssHead}) object of the note to be posted.}

\item{csColorBody}{(\link{character}) A String or Text (i.e. \link{character} \link{vector}) value that defines
the text colour for the main text (\code{ssBody}) object of the note to be posted.}

\item{csColorCarat}{(\link{character}) A String or Text (i.e. \link{character} \link{vector}) value that defines
the colour for the leading (starting) symbol or icon (\code{csCarat}) object of
the note to be posted.}

\item{csColorSplit}{(\link{character}) A String or Text (i.e. \link{character} \link{vector}) value that defines
the colour for the middle spacer (\code{csSplit}) object of the note to be posted.}

\item{sbPrePendNL}{(\link{logical}) A Boolean value that specifies whether a new line (blank space or
row) should be added to the START of the note to be posted or not (default ==
\code{FALSE}).}

\item{sbPostPend1NL}{(\link{logical}) A Boolean value that specifies whether a SINGLE new line (blank
space or row) should be added to the END of the note to be posted or not
(default == \code{TRUE}).}

\item{sbPostPend2NLs}{(\link{logical}) A Boolean value that specifies whether a DOUBLE new line (blank
spaces or rows) should be added to the END of the note to be posted or not.
This function argument can be useful for clearly delineating sections within
a custom R Function or R Project output (printed) code (default == \code{FALSE}).}

\item{sbPostAlways}{(\link{logical}) A Boolean value that specifies whether the note should ALWAYS be
posted (printed to the active R Console) or not (default == \code{TRUE}).}

\item{siCallCELN}{(\link{integer}) A Numeric value that specifies the Code Editor Line Number (CELN)
at which this function was called from by its parent function or R Script.}
}
\value{
\itemize{
\item This function prints the supplied text (message) directly to the active R Session Console when
the function argument \code{sbPostAlways} (or either of the R Project DEBUG and VERBOSE trackers:
i.e. \code{RCT_IS_DEBUG_MODE_} and \code{RCT_IS_VERBOSE_MODE_}, respectively) are set to \code{TRUE}.
\item This function also outputs the full (complete) notification message as an invisible function
return value (function result).
}
}
\description{
A Helper Function that standardizes the R Project Information Posting (i.e. User Notification)
Processes. This function was intended to mainly support the MFMR Suite of R Functions, but may
be utilized as a standalone function in other (i.e. 3rd Party) R packages.
}
\examples{
### Easily post notifications to the active R Console as follows ...
library(MFMRutils)   # <- Ensures the `MFMRutils` library is <already> installed locally.


### Example 1: Post a dummy note (default notification) ...
info.post.note()   # <- Run function without <user-specified> inputs to generate a dummy post !!!


### Example 2: Post a user-specified notification ...
info.post.note(
  csCarat = "~>",                               # <- Sets the leading <start> icon to a "~>".
  ssHead = "My-CUST-FUNC",                      # <- Sets the HEADER text of the notification ...
  ssBody = "This is my COOL NOTE -> YaY !!!",   # <- Sets the MAIN <body> text of notification.
  sbShowTail = FALSE                            # <- Hides the trailing <tail> icon object !!!
)


### Example 3: Deactivate the built-in formatting (spaces, font weights and font colours ) ...
info.post.note(
  csCarat = "~>",                               
  ssHead = "My-CUST-FUNC",                      
  ssBody = "This is my COOL NOTE -> YaY !!!",   
  sbShowTail = FALSE,                           
  sbPrintPretty = FALSE   # <- Prints the input text as provided (i.e. without spaces or any 
)                         #    text formatting). You must format your text (as required) prior to
                          #    handing it to the function (when `sbPrintPretty` is set to FALSE).


### Example 4: Post notifications on basis of the DEBUG and VERBOSE R Project trackers ...
RCT_IS_VERBOSE_MODE_ <- TRUE   # <- Set `RCT_IS_VERBOSE_MODE_` to TRUE anywhere in R Project ...
info.post.note(
  csCarat = "~>",                               
  ssHead = "My-CUST-FUNC",                      
  ssBody = "This is my COOL NOTE -> YaY !!!",   
  sbPostAlways = FALSE         # <- The R Project VERBOSE tracker (i.e. `RCT_IS_VERBOSE_MODE_`)
)                              #    overrides the `sbPostAlways` function argument -> this means
                               #    the function WILL PRINT its message to the R Console, even if
                               #    the `sbPostAlways` function argument is set to FALSE, when 
                               #    the VERBOSE tracker is set to a value of TRUE !!!
                               
## Remove VERBOSE Tracker from Global R Environment ...
rm(list = c("RCT_IS_VERBOSE_MODE_"))   # <- Remove VERBOSE Tracker from R Environment !!!


RCT_IS_DEBUG_MODE_ <- TRUE   # <- Set `RCT_IS_DEBUG_MODE_` to TRUE anywhere in an R Project ...
info.post.note(
  csCarat = "~>",
  ssHead = "My-CUST-FUNC",
  ssBody = "This is my COOL NOTE -> YaY !!!",
  sbPostAlways = FALSE       # <- The R Project DEBUG tracker (i.e. `RCT_IS_DEBUG_MODE_`)
)                            #    overrides the `sbPostAlways` function argument -> this means
                             #    the function WILL PRINT its message to the R Console, even if
                             #    the `sbPostAlways` function argument is set to FALSE, when
                             #    the DEBUG tracker is set to a value of TRUE !!!

### NB: DEBUG Tracker also activates the Code Editor Line Number (CELN) section of the 
###    `info.post.note()` function output (CELN is positioned after the HEADER text) !!!
                               
## Remove DEBUG Tracker from Global R Environment ...
rm(list = c("RCT_IS_DEBUG_MODE_"))   # <- Remove DEBUG Tracker from R Environment !!!


}
\seealso{
Other SuiteMFMR INFO Functions: 
\code{\link{info.post.func.self.id}()}
}
\concept{SuiteMFMR INFO Functions}
