% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.tiny.helpers.R
\name{"pkgs.get.lib.info"}
\alias{"pkgs.get.lib.info"}
\alias{pkgs.get.lib.info}
\title{Extract R Package \if{html}{\out{<Library>}} Information}
\usage{
pkgs.get.lib.info(ssPathToFileDESC = "./DESCRIPTION")
}
\description{
A tiny R function that extracts relevant package information from \if{html}{\out{<internal>}}
R Library Attributes.
}
\examples{
### Easily print & assign Date Formats as follows ...
ENVDates$LONGv01   ### -> formats Date to "Saturday, January 16 1982" !!!
ENVDates$LONGv02   ### -> formats Date to "Sat, Jan 16 1982" !!!
ENVDates$LONGv03   ### -> formats Date to "Sat, 16 Jan 1982" !!!

### Use with the direct-access R operator "::" from anywhere ...
MFMRutils::ENVFormats$BOLD   ### -> sets the text font format to BOLD !!!

}
